Dizkord.STYLES — Constant

Regex expressions for split_message to not break Discord formatting.

Dizkord.AbstractEvent — Type

An incoming event sent over the gateway. Also a catch-all event: Handlers defined on this type will execute on all events.

Dizkord.AbstractGuild — Type

A Discord guild (server). Can either be an UnavailableGuild or a Guild.

Dizkord.ActionType — Type

AuditLog action types. More details here.

Dizkord.Activity — Type

A User activity. More details here.

Fields

name           :: String
type           :: Dizkord.ActivityType
url            :: OptionalNullable{String}
timestamps     :: Optional{Dizkord.ActivityTimestamps}
application_id :: Optional{Snowflake}
details        :: OptionalNullable{String}
state          :: OptionalNullable{String}
emoji          :: OptionalNullable{Dizkord.ActivityEmoji}
party          :: Optional{Dizkord.ActivityParty}
assets         :: Optional{Dizkord.ActivityAssets}
secrets        :: Optional{Dizkord.ActivitySecrets}
instance       :: Optional{Bool}
flags          :: Optional{Int}

Dizkord.ActivityAssets — Type

Images and hover text for an Activity. More details here.

Fields

large_image :: Optional{String}
large_text  :: Optional{String}
small_image :: Optional{String}
small_text  :: Optional{String}

Dizkord.ActivityEmoji — Type

Emoji for a custom Activity. More details here.

Fields

name     :: String
id       :: Optional{Snowflake}
animated :: Optional{Bool}

Dizkord.ActivityFlags — Type

Flags which indicate what an Activity payload contains. More details here.

Dizkord.ActivityParty — Type

The current party of an Activity's player. More details here.

Fields

id   :: Optional{String}
size :: Optional{Vector{Int}}

Dizkord.ActivitySecrets — Type

Secrets for Rich Presence joining and spectating of an Activity. More details here.

Fields

join     :: Optional{String}
spectate :: Optional{String}
match    :: Optional{String}

Dizkord.ActivityTimestamps — Type

The start and stop times of an Activity. More details here.

Fields

start :: Optional{DateTime}
stop  :: Optional{DateTime}

Dizkord.ActivityType — Type

An Activity's type. Available values are AT_GAME, AT_STREAMING, AT_LISTENING, AT_WATCHING, and AT_COMPETING. More details here.

Dizkord.ApplicationCommand — Type

An Application Command More details here

Fields

id                  :: OptionalNullable{Snowflake}
type                :: Optional{Dizkord.ApplicationCommandType}
application_id      :: Snowflake
guild_id            :: Optional{Snowflake}
name                :: String
description         :: String
options             :: Optional{Vector{Dizkord.ApplicationCommandOption}}
default_permissions :: Optional{Bool}
version             :: Optional{Snowflake}

Dizkord.ApplicationCommandChoice — Type

Application Command Choice More details here

Fields

name  :: String
value :: Union{Number, String}

Dizkord.ApplicationCommandOption — Type

Application Command Option More details here

Fields

name          :: Optional{String}
type          :: Dizkord.OptionType
value         :: Any
required      :: Optional{Bool}
choices       :: Optional{Vector{Dizkord.ApplicationCommandChoice}}
options       :: Optional{Vector{Dizkord.ApplicationCommandOption}}
channel_types :: Optional{Vector{Dizkord.ChannelTypes}}

Dizkord.Attachment — Type

A Message attachment. More details here.

Fields

id        :: Snowflake
filename  :: String
size      :: Int
url       :: String
proxy_url :: String
height    :: Optional{Int}
width     :: Optional{Int}

Dizkord.AuditLog — Type

An audit log. More details here.

Fields

webhooks          :: Vector{Dizkord.Webhook}
users             :: Vector{Dizkord.User}
audit_log_entries :: Vector{Dizkord.AuditLogEntry}

Dizkord.AuditLogChange — Type

A change item in an AuditLogEntry.

The first type parameter is the type of new_value and old_value. The second is the type of the entity that new_value and old_value belong(ed) to.

More details here.

Fields

new_value :: Optional{T} where T
old_value :: Optional{T} where T
key       :: String
type      :: Type{U} where U

Dizkord.AuditLogEntry — Type

An entry in an AuditLog. More details here.

Fields

target_id   :: Nullable{Snowflake}
changes     :: Optional{Vector{Dizkord.AuditLogChange}}
user_id     :: Snowflake
id          :: Snowflake
action_type :: Dizkord.ActionType
options     :: Optional{Dizkord.AuditLogOptions}
reason      :: Optional{String}

Dizkord.AuditLogOptions — Type

Optional information in an AuditLogEntry. More details here.

Fields

delete_member_days :: Optional{Int}
members_removed    :: Optional{Int}
channel_id         :: Optional{Snowflake}
count              :: Optional{Int}
id                 :: Optional{Snowflake}
type               :: Optional{Dizkord.OverwriteType}
role_name          :: Optional{String}

Dizkord.Ban — Type

A User ban. More details here.

Fields

reason :: Nullable{String}
user   :: Dizkord.User

Dizkord.CacheFilter — Type

CacheFilter(f::Function) -> CacheFilter

Only store value v at key k if f(v) === true (k is always v.id).

Dizkord.CacheForever — Type

CacheForever() -> CacheForever

Store everything and never evict items from the cache.

Dizkord.CacheLRU — Type

CacheLRU(size::Int) -> CacheLRU

Evict the least recently used item from the cache when there are more than size items.

Dizkord.CacheNever — Type

CacheNever() -> CacheNever

Don't store anything in the cache.

Dizkord.CacheStrategy — Type

A method of handling cache insertion and eviction.

Dizkord.CacheTTL — Type

CacheTTL(ttl::Period) -> CacheTTL

Evict items from the cache after ttl has elapsed.

Dizkord.ChannelCreate — Type

Sent when a new DiscordChannel is created.

Fields

channel :: Dizkord.DiscordChannel

Dizkord.ChannelDelete — Type

Sent when a DiscordChannel is deleted.

Fields

channel :: Dizkord.DiscordChannel

Dizkord.ChannelPinsUpdate — Type

Sent when a DiscordChannel's pins are updated.

Fields

channel_id         :: Snowflake
last_pin_timestamp :: Nullable{DateTime}

Dizkord.ChannelType — Type

A DiscordChannel's type. Available values are CT_GUILD_TEXT, CT_DM, CT_GUILD_VOICE, CT_GROUP_DM, CT_GUILD_CATEGORY, CT_GUILD_NEWS and CT_GUILD_STORE.

Dizkord.ChannelUpdate — Type

Sent when a DiscordChannel is updated.

Fields

channel :: Dizkord.DiscordChannel

Dizkord.Client — Type

Client(
    token::String
    application_id::Snowflake
    intents::Int;
    presence::Union{Dict, NamedTuple}=Dict(),
    strategies::Dict{DataType, <:CacheStrategy}=Dict(),
    version::Int=9,
) -> Client

A Discord bot. Clients can connect to the gateway, respond to events, and make REST API calls to perform actions such as sending/deleting messages, kicking/banning users, etc.

Bot Token

A bot token can be acquired by creating a new application here. Make sure not to hardcode the token into your Julia code! Use an environment variable or configuration file instead.

Application ID

The application id for your bot can be found here. Make sure not to hardcode the application id into your Julia code! Use an environment variable or configuration file instead.

Intents

Integer representing intents. More information here

Presence

The presence keyword sets the bot's presence upon connection. It also sets defaults for future calls to set_game. The schema here must be followed.

Cache Control

By default, most data that comes from Discord is cached for later use. However, to avoid memory leakage, not all of it is kept forever. The default setings are to keep everything but Messages, which are deleted after 6 hours, forever. Although the default settings are sufficient for most workloads, you can specify your own strategies per type with the strategies keyword. Keys can be any of the following:

Guild

DiscordChannel

Message

User

Member

Presence

For potential values, see CacheStrategy.

The cache can also be disabled/enabled permanently and temporarily as a whole with enable_cache! and disable_cache!.

API Version

The version keyword chooses the Version of the Discord API to use. Using anything but 9 is not officially supported by the Dizkord.jl developers.

Sharding

Sharding is handled automatically. The number of available processes is the number of shards that are created. See the sharding example for more details.

Dizkord.Connection — Type

A User connection to an external service (Twitch, YouTube, etc.). More details here.

Fields

id           :: String
name         :: String
type         :: String
revoked      :: Bool
integrations :: Vector{Dizkord.Integration}

Dizkord.DiscordChannel — Type

A Discord channel. More details here.

Note: The name Channel is already used, hence the prefix.

Fields

id                    :: Snowflake
type                  :: Dizkord.ChannelType
guild_id              :: Optional{Snowflake}
position              :: Optional{Int}
permission_overwrites :: Optional{Vector{Dizkord.Overwrite}}
name                  :: Optional{String}
topic                 :: OptionalNullable{String}
nsfw                  :: Optional{Bool}
last_message_id       :: OptionalNullable{Snowflake}
bitrate               :: Optional{Int}
user_limit            :: Optional{Int}
rate_limit_per_user   :: Optional{Int}
recipients            :: Optional{Vector{Dizkord.User}}
icon                  :: OptionalNullable{String}
owner_id              :: Optional{Snowflake}
application_id        :: Optional{Snowflake}
parent_id             :: OptionalNullable{Snowflake}
last_pin_timestamp    :: OptionalNullable{DateTime}

Dizkord.Embed — Type

A Message embed. More details here.

Fields

title       :: Optional{String}
type        :: Optional{String}
description :: Optional{String}
url         :: Optional{String}
timestamp   :: Optional{DateTime}
color       :: Optional{Int}
footer      :: Optional{Dizkord.EmbedFooter}
image       :: Optional{Dizkord.EmbedImage}
thumbnail   :: Optional{Dizkord.EmbedThumbnail}
video       :: Optional{Dizkord.EmbedVideo}
provider    :: Optional{Dizkord.EmbedProvider}
author      :: Optional{Dizkord.EmbedAuthor}
fields      :: Optional{Vector{Dizkord.EmbedField}}

Dizkord.EmbedAuthor — Type

An Embed's author information. More details here.

Fields

name           :: Optional{String}
url            :: Optional{String}
icon_url       :: Optional{String}
proxy_icon_url :: Optional{String}

Dizkord.EmbedField — Type

An Embed field. More details here.

Fields

name   :: String
value  :: String
inline :: Optional{Bool}

Dizkord.EmbedFooter — Type

An Embed's footer information. More details here.

Fields

text           :: String
icon_url       :: Optional{String}
proxy_icon_url :: Optional{String}

Dizkord.EmbedImage — Type

An Embed's image information. More details here.

Fields

url       :: Optional{String}
proxy_url :: Optional{String}
height    :: Optional{Int}
width     :: Optional{Int}

Dizkord.EmbedProvider — Type

An Embed's provider information. More details here.

Fields

name :: Optional{String}
url  :: OptionalNullable{String}

Dizkord.EmbedThumbnail — Type

An Embed's thumbnail image information. More details here.

Fields

url       :: Optional{String}
proxy_url :: Optional{String}
height    :: Optional{Int}
width     :: Optional{Int}

Dizkord.EmbedVideo — Type

An Embed's video information. More details here.

Fields

url    :: Optional{String}
height :: Optional{Int}
width  :: Optional{Int}

Dizkord.Emoji — Type

An emoji. More details here.

Fields

id             :: Nullable{Snowflake}
name           :: String
roles          :: Optional{Vector{Snowflake}}
user           :: Optional{Dizkord.User}
require_colons :: Optional{Bool}
managed        :: Optional{Bool}
animated       :: Optional{Bool}

Dizkord.ExplicitContentFilterLevel — Type

A Guild's explicit content filter level. More details here.

Dizkord.FallbackEvent — Type

A type for defining handlers on any events which would otherwise have no non-default handler. Handlers for this type must accept an AbstractEvent.

Dizkord.Guild — Type

A Discord guild (server). More details here.

Fields

id                            :: Snowflake
name                          :: String
icon                          :: Nullable{String}
splash                        :: OptionalNullable{String}
owner                         :: Optional{Bool}
owner_id                      :: Optional{Snowflake}
permissions                   :: Optional{String}
region                        :: Optional{String}
afk_channel_id                :: OptionalNullable{Snowflake}
afk_timeout                   :: Optional{Int}
embed_enabled                 :: Optional{Bool}
embed_channel_id              :: OptionalNullable{Snowflake}
verification_level            :: Optional{Dizkord.VerificationLevel}
default_message_notifications :: Optional{Dizkord.MessageNotificationLevel}
explicit_content_filter       :: Optional{Dizkord.ExplicitContentFilterLevel}
roles                         :: Optional{Vector{Dizkord.Role}}
emojis                        :: Optional{Vector{Dizkord.Emoji}}
features                      :: Optional{Vector{String}}
mfa_level                     :: Optional{Dizkord.MFALevel}
application_id                :: OptionalNullable{Snowflake}
widget_enabled                :: Optional{Bool}
widget_channel_id             :: OptionalNullable{Snowflake}
system_channel_id             :: OptionalNullable{Snowflake}
joined_at                     :: Optional{DateTime}
large                         :: Optional{Bool}
unavailable                   :: Optional{Bool}
member_count                  :: Optional{Int}
max_members                   :: Optional{Int}
voice_states                  :: Optional{Vector{Dizkord.VoiceState}}
members                       :: Optional{Vector{Dizkord.Member}}
channels                      :: Optional{Vector{Dizkord.DiscordChannel}}
presences                     :: Optional{Vector{Dizkord.Presence}}
max_presences                 :: OptionalNullable{Int}
vanity_url_code               :: OptionalNullable{String}
description                   :: OptionalNullable{String}
banner                        :: OptionalNullable{String}

Dizkord.GuildBanAdd — Type

Sent when a User is banned from a Guild.

Fields

guild_id :: Snowflake
user     :: Dizkord.User

Dizkord.GuildBanRemove — Type

Sent when a User is unbanned from a Guild.

Fields

guild_id :: Snowflake
user     :: Dizkord.User

Dizkord.GuildCreate — Type

Sent when a new Guild is created.

Fields

guild :: Dizkord.Guild

Dizkord.GuildDelete — Type

Sent when a Guild is deleted.

Fields

guild :: Dizkord.AbstractGuild

Dizkord.GuildEmbed — Type

A Guild embed. More details here.

Fields

enabled    :: Bool
channel_id :: Nullable{Snowflake}

Dizkord.GuildEmojisUpdate — Type

Sent when a Guild has its Emojis updated.

Fields

guild_id :: Snowflake
emojis   :: Vector{Dizkord.Emoji}

Dizkord.GuildIntegrationsUpdate — Type

Sent when a Guild has its Integrations updated.

Fields

guild_id :: Snowflake

Dizkord.GuildMemberAdd — Type

Sent when a Member is added to a Guild.

Fields

guild_id :: Snowflake
member   :: Dizkord.Member

Dizkord.GuildMemberRemove — Type

Sent when a Member is removed from a Guild.

Fields

guild_id :: Snowflake
user     :: Dizkord.User

Dizkord.GuildMemberUpdate — Type

Sent when a Member is updated in a Guild.

Fields

guild_id :: Snowflake
roles    :: Vector{Snowflake}
user     :: Dizkord.User
nick     :: Nullable{String}

Dizkord.GuildMembersChunk — Type

Sent when the Client requests guild members with request_guild_members.

Fields

guild_id :: Snowflake
members  :: Vector{Dizkord.Member}

Dizkord.GuildRoleCreate — Type

Sent when a new Role is created in a Guild.

Fields

guild_id :: Snowflake
role     :: Dizkord.Role

Dizkord.GuildRoleDelete — Type

Sent when a Role is deleted from a Guild.

Fields

guild_id :: Snowflake
role_id  :: Snowflake

Dizkord.GuildRoleUpdate — Type

Sent when a Role is updated in a Guild.

Fields

guild_id :: Snowflake
role     :: Dizkord.Role

Dizkord.GuildUpdate — Type

Sent when a Guild is updated.

Fields

guild :: Dizkord.Guild

Dizkord.Integration — Type

A Guild integration. More details here.

Fields

id                  :: Snowflake
name                :: String
type                :: String
enabled             :: Bool
syncing             :: Bool
role_id             :: Snowflake
expire_behaviour    :: Int
expire_grace_period :: Int
user                :: Dizkord.User
account             :: Dizkord.IntegrationAccount
synced_at           :: DateTime

Dizkord.IntegrationAccount — Type

An Integration account. More details here.

Fields

id   :: String
name :: String

Dizkord.Interaction — Type

An interaction. More details here.

Fields

id             :: Nullable{Snowflake}
application_id :: Nullable{Snowflake}
type           :: Dizkord.InteractionType
data           :: OptionalNullable{Dizkord.InteractionData}
guild_id       :: Optional{Snowflake}
channel_id     :: Optional{Snowflake}
member         :: Optional{Dizkord.Member}
user           :: Optional{Dizkord.User}
token          :: String
version        :: Optional{Int}
message        :: Optional{Dizkord.Message}

Dizkord.InteractionCreate — Type

Interaction Create gateway event

Fields

int :: Dizkord.Interaction

Dizkord.InteractionData — Type

Data for an interaction. More details here.

Fields

id             :: Nullable{Snowflake}
name           :: String
type           :: Dizkord.ApplicationCommandType
resolved       :: Optional{Dizkord.ResolvedData}
options        :: Optional{Vector{Dizkord.ApplicationCommandOption}}
custom_id      :: OptionalNullable{String}
component_type :: OptionalNullable{Dizkord.ComponentType}
values         :: Optional{Vector{String}}
target_id      :: Optional{Snowflake}

Dizkord.Invite — Type

An invite to a Guild. More details here.

Fields

code                       :: String
guild                      :: Optional{Dizkord.Guild}
channel                    :: Dizkord.DiscordChannel
approximate_presence_cound :: Optional{Int}
approximate_member_count   :: Optional{Int}

Dizkord.InviteMetadata — Type

Metadata for an Invite. More details here.

Fields

inviter    :: Dizkord.User
uses       :: Int
max_uses   :: Int
max_age    :: Int
temporary  :: Bool
created_at :: DateTime
revoked    :: Bool

Dizkord.MFALevel — Type

A Guild's MFA level. More details here.

Dizkord.Member — Type

A Guild member. More details here.

Fields

user          :: Optional{Dizkord.User}
nick          :: OptionalNullable{String}
roles         :: Vector{Snowflake}
joined_at     :: DateTime
premium_since :: OptionalNullable{DateTime}
deaf          :: Bool
mute          :: Bool

Dizkord.Message — Type

A message sent to a DiscordChannel. More details here.

Fields

id               :: Snowflake
channel_id       :: Snowflake
guild_id         :: Optional{Snowflake}
author           :: Optional{Dizkord.User}
member           :: Optional{Dizkord.Member}
content          :: Optional{String}
timestamp        :: Optional{DateTime}
edited_timestamp :: OptionalNullable{DateTime}
tts              :: Optional{Bool}
mention_everyone :: Optional{Bool}
mentions         :: Optional{Vector{Dizkord.User}}
mention_roles    :: Optional{Vector{Snowflake}}
attachments      :: Optional{Vector{Dizkord.Attachment}}
embeds           :: Optional{Vector{Dizkord.Embed}}
reactions        :: Optional{Vector{Dizkord.Reaction}}
nonce            :: OptionalNullable{Snowflake}
pinned           :: Optional{Bool}
webhook_id       :: Optional{Snowflake}
type             :: Optional{Dizkord.MessageType}
activity         :: Optional{Dizkord.MessageActivity}
application      :: Optional{Dizkord.MessageApplication}

Dizkord.MessageActivity — Type

A Message activity. More details here.

Fields

type     :: Dizkord.MessageActivityType
party_id :: Optional{String}

Dizkord.MessageActivityType — Type

A Message's activity type. More details here.

Dizkord.MessageApplication — Type

A Rich Presence Message's application information. More details here.

Fields

id          :: Snowflake
cover_image :: Optional{String}
description :: String
icon        :: String
name        :: String

Dizkord.MessageCreate — Type

Sent when a Message is sent.

Fields

message :: Dizkord.Message

Dizkord.MessageDelete — Type

Sent when a Message is deleted.

Fields

id         :: Snowflake
channel_id :: Snowflake
guild_id   :: Optional{Snowflake}

Dizkord.MessageDeleteBulk — Type

Sent when multiple Messages are deleted in bulk.

Fields

ids        :: Vector{Snowflake}
channel_id :: Snowflake
guild_id   :: Optional{Snowflake}

Dizkord.MessageNotificationLevel — Type

A Guild's default message notification level. More details here.

Dizkord.MessageReactionAdd — Type

Sent when a Reaction is added to a Message.

Fields

user_id    :: Snowflake
channel_id :: Snowflake
message_id :: Snowflake
guild_id   :: Optional{Snowflake}
emoji      :: Dizkord.Emoji

Dizkord.MessageReactionRemove — Type

Sent when a Reaction is removed from a Message.

Fields

user_id    :: Snowflake
channel_id :: Snowflake
message_id :: Snowflake
guild_id   :: Optional{Snowflake}
emoji      :: Dizkord.Emoji

Dizkord.MessageReactionRemoveAll — Type

Sent when all Reactions are removed from a Message.

Fields

channel_id :: Snowflake
message_id :: Snowflake
guild_id   :: Optional{Snowflake}

Dizkord.MessageType — Type

A Message's type. More details here.

Dizkord.MessageUpdate — Type

Sent when a Message is updated.

Fields

message :: Dizkord.Message

Dizkord.OnMessageCreate — Type

Handler for a Message Create event

Dizkord.Overwrite — Type

A permission overwrite. More details here.

Fields

id    :: Snowflake
type  :: Dizkord.OverwriteType
allow :: String
deny  :: String

Dizkord.OverwriteType — Type

An Overwrite's type. More details here.

Dizkord.Permission — Type

Bitwise permission flags. More details here.

Dizkord.Presence — Type

A User's presence. More details here.

Fields

user       :: Dizkord.User
roles      :: Optional{Vector{Snowflake}}
game       :: Nullable{Dizkord.Activity}
guild_id   :: Optional{Snowflake}
status     :: Dizkord.PresenceStatus
activities :: Vector{Dizkord.Activity}

Dizkord.PresenceStatus — Type

A User's status sent in a Presence. Available values are PS_IDLE, PS_DND, PS_ONLINE, and PS_OFFLINE. More details here.

Dizkord.PresenceUpdate — Type

Sent when a User's Presence is updated.

Fields

presence :: Dizkord.Presence

Dizkord.Reaction — Type

A Message reaction. More details here.

Fields

count :: Int
me    :: Bool
emoji :: Dizkord.Emoji

Dizkord.Ready — Type

Sent when the Client has authenticated, and contains the initial state.

Fields

v                :: Int
user             :: Dizkord.User
private_channels :: Vector{Dizkord.DiscordChannel}
guilds           :: Vector{Dizkord.UnavailableGuild}
session_id       :: String
_trace           :: Vector{String}

Dizkord.Response — Type

A wrapper around a response from the REST API. Every function which wraps a Discord REST API endpoint returns a Future which will contain a value of this type. To retrieve the Response from the Future, use fetch. See also: fetchval.

Fields

val::Nullable{T}: The object contained in the HTTP response. For example, for a call to get_channel_message, this value will be a Message.

ok::Bool: The state of the request. If true, then it is safe to access val.

http_response::Nullable{HTTP.Messages.Response}: The underlying HTTP response, if a request was made.

exception::Nullable{Exception}: The caught exception, if one is thrown.

Examples

Multiple API calls which immediately return Futures and can be awaited:

futures = map(i -> create_message(c, channel_id; content=string(i)), 1:10);
other_work_here()
resps = fetch.(futures)

Skipping error checks and returning the value directly:

guild = fetchval(create_guild(c; name="foo"))

Dizkord.Resumed — Type

Sent when a Client resumes its connection.

Fields

_trace :: Vector{String}

Dizkord.Role — Type

A User role. More details here.

Fields

id          :: Snowflake
name        :: String
color       :: Optional{Int}
hoist       :: Optional{Bool}
position    :: Optional{Int}
permissions :: Optional{String}
managed     :: Optional{Bool}
mentionable :: Optional{Bool}

Dizkord.TypingStart — Type

Sent when a User begins typing.

Fields

channel_id :: Snowflake
guild_id   :: Optional{Snowflake}
user_id    :: Snowflake
timestamp  :: DateTime

Dizkord.UnavailableGuild — Type

An unavailable Discord guild (server). More details here.

Fields

id          :: Snowflake
unavailable :: Optional{Bool}

Dizkord.UnknownEvent — Type

An unknown event. When an event can't be parsed, due to an unknown type or any other error, it will appear as an UnknownEvent. The fields follow the schema defined here.

Dizkord.User — Type

A Discord user. More details here.

Fields

id            :: Snowflake
username      :: Optional{String}
discriminator :: Optional{String}
avatar        :: OptionalNullable{String}
bot           :: Optional{Bool}
mfa_enabled   :: Optional{Bool}
locale        :: Optional{String}
verified      :: Optional{Bool}
email         :: OptionalNullable{String}

Dizkord.UserUpdate — Type

Sent when a User's details are updated.

Fields

user :: Dizkord.User

Dizkord.VerificationLevel — Type

A Guild's verification level. More details here.

Dizkord.VoiceRegion — Type

A region for a Guild's voice server. More details here.

Fields

id         :: String
name       :: String
vip        :: Bool
optimal    :: Bool
deprecated :: Bool
custom     :: Bool

Dizkord.VoiceServerUpdate — Type

Sent when a Guild's voice server is updated.

Fields

token    :: String
guild_id :: Snowflake
endpoint :: String

Dizkord.VoiceState — Type

A User's voice connection status. More details here.

Fields

guild_id   :: Optional{Snowflake}
channel_id :: Nullable{Snowflake}
user_id    :: Snowflake
member     :: Optional{Dizkord.Member}
session_id :: String
deaf       :: Bool
mute       :: Bool
self_deaf  :: Bool
self_mute  :: Bool
suppress   :: Bool

Dizkord.VoiceStateUpdate — Type

Sent when a User updates their VoiceState.

Fields

state :: Dizkord.VoiceState

Dizkord.Webhook — Type

A Webhook. More details here.

Fields

id         :: Snowflake
guild_id   :: Optional{Snowflake}
channel_id :: Snowflake
user       :: Optional{Dizkord.User}
name       :: Nullable{String}
avatar     :: Nullable{String}
token      :: Optional{String}

Dizkord.WebhooksUpdate — Type

Sent when a DiscordChannel's Webhooks are updated.

Fields

guild_id   :: Snowflake
channel_id :: Snowflake

Base.close — Method

close(c::Client)

Disconnect the Client from the gateway.

Base.isopen — Method

isopen(c::Client) -> Bool

Determine whether the Client is connected to the gateway.

Base.open — Method

open(c::Client; delay::Period=Second(7))

Connect a Client to the Discord gateway.

The delay keyword is the time between shards connecting. It can be increased from its default if you are using multiple shards and frequently experiencing invalid sessions upon connection.

Base.wait — Method

wait(c::Client)

Wait for an open Client to close.

Dizkord.add_guild_member — Method

add_guild_member(c::Client; kwargs...) -> Member

Add a User to a Guild. More details here.

Dizkord.add_guild_member_role — Method

add_guild_member_role(c::Client, guild::Integer, user::Integer, role::Integer)

Add a Role to a Member.

Dizkord.add_pinned_channel_message — Method

add_pinned_channel_message(c::Client, channel::Integer, message::Integer)

Pin a Message in a DiscordChannel.

Dizkord.begin_guild_prune — Method

begin_guild_prune(c::Client, guild::Integer; kwargs...) -> Dict

Begin pruning Members from a Guild. More details here.

Dizkord.bulk_delete_messages — Method

bulk_delete_messages(c::Client, channel::Integer; kwargs...)

Delete multiple Messages. More details here.

Dizkord.command! — Method

command!(
    f::Function
    c::Client
    name::AbstractString
    description::AbstractString;
    kwargs...
)

Adds a handler for INTERACTION CREATE gateway events where the InteractionData's name field matches name. Adds this command to c.commands or c.guild_commands based on the presence of guild. The f parameter signature should be:

    (ctx::OnInteractionCreateContext) -> Any 

Dizkord.context — Method

context(::Type{<:AbstractHandler}, data) -> AbstractContext

Generates the context for a Handler based on the given data.

Dizkord.create — Function

create(c::Client, ::Type{T}, args...; kwargs...) -> Future{Response}

Create, add, send, etc.

Examples

Sending a Message:

create(c, Message, channel; content="foo")

Creating a new DiscordChannel:

create(c, DiscordChannel, guild; name="bar")

Banning a Member:

create(c, Ban, guild, member; reason="baz")

Dizkord.create_channel_invite — Method

create_channel_invite(c::Client, channel::Integer; kwargs...) -> Invite

Create an Invite to a DiscordChannel. More details here.

Dizkord.create_dm — Method

create_dm(c::Client; kwargs...) -> DiscordChannel

Create a DM DiscordChannel. More details here.

Dizkord.create_guild — Method

create_guild(c::Client; kwargs...) -> Guild

Create a Guild. More details here.

Dizkord.create_guild_ban — Method

create_guild_ban(c::Client, guild::Integer, user::Integer; kwargs...)

Ban a Member from a Guild. More details here.

Dizkord.create_guild_channel — Method

create_guild_channel(c::Client, guild::Integer; kwargs...) -> DiscordChannel

Create a DiscordChannel in a Guild. More details here.

Dizkord.create_guild_emoji — Method

create_guild_emoji(c::Client, guild::Integer; kwargs...) -> Emoji

Create an Emoji in a Guild. More details here.

Dizkord.create_guild_integration — Method

create_guild_integration(c::Client, guild::Integer; kwargs...)

Create/attach an Integration to a Guild. More details here.

Dizkord.create_guild_role — Method

create_guild_role(c::Client, guild::Integer; kwargs) -> Role

Create a Role in a Guild. More details here.

Dizkord.create_message — Method

create_message(c::Client, channel::Integer; kwargs...) -> Message

Send a Message to a DiscordChannel. More details here.

Dizkord.create_reaction — Method

create_reaction(c::Client, channel::Integer, message::Integer, emoji::StringOrChar)

React to a Message. If emoji is a custom Emoji, it should be formatted "name:id".

Dizkord.create_webhook — Method

create_webhook(c::Client, channel::Integer; kwargs...) -> Webhook

Create a Webhook in a DiscordChannel. More details here.

Dizkord.delete — Function

delete(c::Client, x::T, args...) -> Future{Response}

Delete, remove, discard, etc.

Examples

Kicking a Member:

delete(c, member)

Unbanning a Member:

delete(c, ban, guild)

Deleting all Reactions from a Message (note: this is the only update/delete method which takes a type parameter):

delete(c, Reaction, message)

Dizkord.delete_all_reactions — Method

delete_all_reactions(c::Client, channel::Integer, message::Integer)

Delete all reactions from a Message.

Dizkord.delete_channel — Method

delete_channel(c::Client, channel::Integer) -> DiscordChannel

Delete a DiscordChannel.

Dizkord.delete_channel_permission — Method

delete_channel_permission(c::Client, channel::Integer, overwrite::Integer)

Delete an Overwrite from a DiscordChannel.

Dizkord.delete_guild — Method

delete_guild(c::Client, guild::Integer)

Delete a Guild.

Dizkord.delete_guild_emoji — Method

delete_guild_emoji(c::Client, guild::Integer, emoji::Integer)

Delete an Emoji from a Guild.

Dizkord.delete_guild_integration — Method

delete_guild_integration(c::Client, guild::Integer, integration::Integer)

Delete an Integration from a Guild.

Dizkord.delete_guild_role — Method

delete_guild_role(c::Client, guild::Integer, role::Integer)

Delete a Role from a Guild.

Dizkord.delete_invite — Method

delete_invite(c::Client, invite::AbstractString) -> Invite

Delete an Invite to a Guild.

Dizkord.delete_message — Method

delete_message(c::Client, channel::Integer, message::Integer)

Delete a Message.

Dizkord.delete_own_reaction — Method

delete_own_reaction(c::Client, channel::Integer, message::Integer, emoji::StringOrChar)

Delete the Client user's reaction to a Message.

Dizkord.delete_pinned_channel_message — Method

delete_pinned_channel_message(c::Client, channel::Integer, message::Integer)

Unpin a Message from a DiscordChannel.

Dizkord.delete_user_reaction — Method

delete_user_reaction(
    c::Client,
    channel::Integer,
    message::Integer,
    emoji::StringOrChar,
    user::Integer,
)

Delete a User's reaction to a Message.

Dizkord.delete_webhook — Method

delete_webhook(c::Client, webhook::Integer)

Delete a Webhook.

Dizkord.delete_webhook_with_token — Method

delete_webhook_with_token(c::Client, webhook::Integer, token::AbstractString)

Delete a Webhook with a token.

Dizkord.disable_cache! — Method

disable_cache!(c::Client)
disable_cache!(f::Function, c::Client)

Disable the cache. do syntax is also accepted.

Dizkord.edit_channel_permissions — Method

edit_channel_permissions(
    c::Client,
    channel::Integer,
    overwrite::Integer;
    kwargs...,
)

Edit permissions for a DiscordChannel. More details here.

Dizkord.edit_message — Method

edit_message(c::Client, channel::Integer, message::Integer; kwargs...) -> Message

Edit a Message. More details here.

Dizkord.enable_cache! — Method

enable_cache!(c::Client)
enable_cache!(f::Function c::Client)

Enable the cache. do syntax is also accepted.

Dizkord.execute_github_compatible_webhook — Method

execute_github_compatible_webhook(
    c::Client,
    webhook::Integer,
    token::AbstractString;
    wait::Bool=true,
    kwargs...,
)

Execute a Github Webhook. More details here.

Dizkord.execute_slack_compatible_webhook — Method

execute_slack_compatible_webhook(
    c::Client,
    webhook::Integer,
    token::AbstractString;
    wait::Bool=true,
    kwargs...,
)

Execute a Slack Webhook. More details here.

Dizkord.execute_webhook — Method

execute_webhook(
    c::Client,
    webhook::Integer,
    token::AbstractString;
    wait::Bool=false,
    kwargs...,
) -> Message

Execute a Webhook. If wait is not set, no Message is returned. More details here.

Dizkord.fetchval — Method

fetchval(f::Future{Response{T}}) -> Nullable{T}

Shortcut for fetch(f).val: Fetch a Response and return its value. Note that there are no guarantees about the response's success and the value being returned, and it discards context that can be useful for debugging, such as HTTP responses and caught exceptions.

Dizkord.filter_ranges — Method

filter_ranges(u::Vector{UnitRange{Int}})

Filter a list of ranges, discarding ranges included in other ranges from the list.

Example

julia> Dizkord.filter_ranges([1:5, 3:8, 1:20, 2:16, 10:70, 25:60, 5:35, 50:90, 10:70])
4-element Vector{UnitRange{Int64}}:
 1:20
 5:35
 50:90
 10:70

Dizkord.get_channel — Method

get_channel(c::Client, channel::Integer) -> DiscordChannel

Get a DiscordChannel.

Dizkord.get_channel_invites — Method

get_channel_invites(c::Client, channel::Integer) -> Vector{Invite}

Get the Invites for a DiscordChannel.

Dizkord.get_channel_message — Method

get_channel_message(c::Client, channel::Integer, message::Integer) -> Message

Get a Message from a DiscordChannel.

Dizkord.get_channel_messages — Method

get_channel_messages(c::Client, channel::Integer; kwargs...) -> Vector{Message}

Get a list of Messages from a DiscordChannel. More details here.

Dizkord.get_channel_webhooks — Method

get_channel_webhooks(c::Client, channel::Integer) -> Vector{Webhook}

Get a list of Webhooks in a DiscordChannel.

Dizkord.get_current_user — Method

get_current_user(c::Client) -> User

Get the Client User.

Dizkord.get_current_user_guilds — Method

get_user_guilds(c::Client; kwargs...) -> Vector{Guild}

Get a list of Guilds the Client User is a member of. More details here.

Dizkord.get_guild — Method

get_guild(c::Client, guild::Integer) -> Guild

Get a Guild.

Dizkord.get_guild_audit_log — Method

get_guild_audit_log(c::Client, guild::Integer; kwargs...) -> AuditLog

Get a Guild's AuditLog. More details here.

Dizkord.get_guild_ban — Method

get_ban(c::Client, guild::Integer,  user::Integer) -> Ban

Get a Ban in a Guild.

Dizkord.get_guild_bans — Method

get_guild_bans(c::Client, guild::Integer) -> Vector{Ban}

Get a list of Bans in a Guild.

Dizkord.get_guild_channels — Method

get_guild_channels(c::Client, guild::Integer) -> Vector{DiscordChannel}

Get the DiscordChannels in a Guild.

Dizkord.get_guild_embed — Method

get_guild_embed(c::Client, guild::Integer) -> GuildEmbed

Get a Guild's GuildEmbed.

Dizkord.get_guild_emoji — Method

get_guild_emoji(c::Client, guild::Integer, emoji::Integer) -> Emoji

Get an Emoji in a Guild.

Dizkord.get_guild_integrations — Method

get_guild_integrations(c::Client, guild::Integer) -> Vector{Integration}

Get a list of Integrations for a Guild.

Dizkord.get_guild_invites — Method

get_guild_invites(c::Client, guild::Integer) -> Vector{Invite}

Get a list of Invites to a Guild.

Dizkord.get_guild_member — Method

get_guild_member(c::Client, guild::Integer, user::Integer) -> Member

Get a Member in a Guild.

Dizkord.get_guild_prune_count — Method

get_guild_prune_count(c::Client, guild::Integer; kwargs...) -> Dict

Get the number of Members that would be removed from a Guild in a prune. More details here.

Dizkord.get_guild_roles — Method

get_guild_roles(c::Client, guild::Integer) -> Vector{Role}

Get a Guild's Roles.

Dizkord.get_guild_voice_regions — Method

get_guild_voice_regions(c::Client, guild::Integer) -> Vector{VoiceRegion}

Get a list of VoiceRegions for the Guild.

Dizkord.get_guild_webhooks — Method

get_guild_webhooks(c::Client, guild::Integer) -> Vector{Webhook}

Get a list of Webhooks in a Guild.

Dizkord.get_guild_widget_image — Method

get_guild_widget_image(c::Client, guild::Integer; kwargs...) -> Vector{UInt8}

Get a Guild's widget image in PNG format. More details here.

Dizkord.get_invite — Method

get_invite(c::Client, invite::AbstractString; kwargs...} -> Invite

Get an Invite to a Guild. More details here.

Dizkord.get_pinned_messages — Method

get_pinned_messages(c::Client, channel::Integer) -> Vector{Message}

Get the pinned Messages in a DiscordChannel.

Dizkord.get_reactions — Method

get_reactions(
    c::Client,
    channel::Integer,
    message::Integer,
    emoji::StringOrChar,
) -> Vector{User}

Get the Users who reacted to a Message with an Emoji.

Dizkord.get_user — Method

get_user(c::Client, user::Integer) -> User

Get a User.

Dizkord.get_vanity_url — Method

get_vanity_url(c::Client, guild::Integer) -> Invite

Get a Guild's vanity URL, if it supports that feature.

Dizkord.get_webhook — Method

get_webhook(c::Client, webhook::Integer) -> Webhook

Get a Webhook.

Dizkord.get_webhook_with_token — Method

get_webhook_with_token(c::Client, webhook::Integer, token::AbstractString) -> Webhook

Get a Webhook with a token.

Dizkord.handle — Method

Calls handle with a list of all found handlers.

Dizkord.handle — Method

Calls handle with the handler type symbol based on the event type.

Dizkord.handle — Method

handle(
    c::Client
    handlers,
    data
)

Determines whether handlers are appropriate to call and calls them if so. Creates an AbstractContext based on the event using the data provided and passes it to the handler.

Dizkord.has_permission — Method

has_permission(perms::Integer, perm::Permission) -> Bool

Determine whether a bitwise OR of permissions contains one Permission.

Examples

julia> has_permission(0x0420, PERM_VIEW_CHANNEL)
true

julia> has_permission(0x0420, PERM_ADMINISTRATOR)
false

julia> has_permission(0x0008, PERM_MANAGE_ROLES)
true

Dizkord.heartbeat_ping — Method

heartbeat_ping(c::Client) -> Nullable{Period}

Get the Client's ping time to the gateway. If the client is not connected, or no heartbeats have been sent/acknowledged, nothing is returned.

Dizkord.leave_guild — Method

leave_guild(c::Client, guild::Integer)

Leave a Guild.

Dizkord.list_guild_emojis — Method

list_guild_emojis(c::Client, guild::Integer) -> Vector{Emoji}

Get the Emojis in a Guild.

Dizkord.list_guild_members — Method

list_guild_members(c::Client, guild::Integer; kwargs...) -> Vector{Member}

Get a list of Members in a Guild. More details here.

Dizkord.list_voice_regions — Method

list_voice_regions(c::Client) -> Vector{VoiceRegion}

Get a list of the VoiceRegions that can be used when creating Guilds.

Dizkord.me — Method

me(c::Client) -> Nullable{User}

Get the Client's bot user.

Dizkord.modify_channel — Method

modify_channel(c::Client, channel::Integer; kwargs...) -> DiscordChannel

Modify a DiscordChannel. More details here.

Dizkord.modify_current_user — Method

modify_current_user(c::Client; kwargs...) -> User

Modify the Client User. More details here.

Dizkord.modify_current_user_nick — Method

modify_current_user_nick(c::Client, guild::Intger; kwargs...) -> String

Modify the Client user's nickname in a Guild. More details here.

Dizkord.modify_guild — Method

modify_guild(c::Client, guild::Integer; kwargs...) -> Guild

Edit a Guild. More details here.

Dizkord.modify_guild_channel_positions — Method

modify_guild_channel_positions(c::Client, guild::Integer, positions...)

Modify the positions of DiscordChannels in a Guild. More details here.

Dizkord.modify_guild_embed — Method

modify_guild_embed(c::Client, guild::Integer; kwargs...) -> GuildEmbed

Modify a Guild's GuildEmbed. More details here.

Dizkord.modify_guild_emoji — Method

modify_guild_emoji(c::Client, guild::Integer, emoji::Integer; kwargs...) -> Emoji

Edit an Emoji in a Guild. More details here.

Dizkord.modify_guild_integration — Method

modify_guild_integration(c::Client, guild::Integer, integration::Integer; kwargs...)

Modify an Integration in a Guild. More details here.

Dizkord.modify_guild_member — Method

modify_guild__member(c::Client, guild::Integer, user::Integer; kwargs...)

Modify a Member in a Guild. More details here.

Dizkord.modify_guild_role — Method

modify_guild_role(c::Client, guild::Integer, role::Integer; kwargs) -> Role

Modify a Role in a Guild. More details here.

Dizkord.modify_guild_role_positions — Method

modify_guild_role_positions(c::Client, guild::Integer, positions...) -> Vector{Role}

Modify the positions of Roles in a Guild. More details here.

Dizkord.modify_webhook — Method

modify_webhook(c::Client, webhook::Integer; kwargs...) -> Webhook

Modify a Webhook. More details here.

Dizkord.modify_webhook_with_token — Method

modify_webhook_with_token(
    c::Client,
    webhook::Integer,
    token::AbstractString;
    kwargs...,
) -> Webhook

Modify a Webhook with a token. More details here.

Dizkord.on_message! — Method

on_message!(
    f::Function
    c::Client
)

Adds a handler for the MESSAGE_CREATE gateway event. The f parameter's signature should be:

    (ctx::OnMessageContext) -> Any 

Dizkord.on_ready! — Method

on_ready!(
    f::Function
    c::Client
)

Adds a handler for the READY gateway event. The f parameter signature should be:

    (ctx::OnReadyContext) -> Any 

Dizkord.permissions_in — Method

permissions_in(m::Member, g::Guild, ch::DiscordChannel) -> Int64

Compute a Member's Permissions in a DiscordChannel.

Dizkord.plaintext — Method

plaintext(m::Message) -> String
plaintext(c::Client, m::Message) -> String

Get the Message contents with any User mentions replaced with their plaintext. If a Client is provided, DiscordChannels Role are also replaced. However, only channels and roles stored in state are replaced; no API requests are made.

Dizkord.remove_guild_ban — Method

remove_guild_ban(c::Client, guild::Integer, user::Integer)

Unban a User from a Guild.

Dizkord.remove_guild_member — Method

remove_guild_member(c::Client, guild::Integer, user::Integer)

Kick a Member from a Guild.

Dizkord.remove_guild_member_role — Method

remove_guild_member_role(c::Client, guild::Integer, user::Integer, role::Integer)

Remove a Role from a Member.

Dizkord.reply — Method

reply(
    c::Client,
    m::Message,
    content::Union{AbstractString, AbstractDict, NamedTuple, Embed};
    at::Bool=false,
) -> Future{Response}

Reply (send a message to the same DiscordChannel) to a Message. If at is set, then the message is prefixed with the sender's mention.

Dizkord.request_guild_members — Method

request_guild_members(
    c::Client,
    guilds::Union{Integer, Vector{<:Integer};
    query::AbstractString="",
    limit::Int=0,
) -> Bool

Request offline guild members of one or more Guilds. GuildMembersChunk events are sent by the gateway in response. More details here.

Dizkord.retrieve — Function

retrieve(c::Client, ::Type{T}, args...; kwargs...) -> Future{Response}

Retrieve, get, list, etc.

Examples

Getting the Client's User:

retrieve(c, User)

Getting a Guild's DiscordChannels:

retrieve(c, DiscordChannel, guild)

Getting an Invite to a Guild by code:

retrieve(c, Invite, "abcdef")

Dizkord.set_game — Method

set_game(
    c::Client,
    game::AbstractString;
    type::Union{ActivityType, Int}=AT_GAME,
    since::Nullable{Int}=c.presence["since"],
    status::Union{PresenceStatus, AbstractString}=c.presence["status"],
    afk::Bool=c.presence["afk"],
    kwargs...,
) -> Bool

Shortcut for update_status to set the Client's Activity. Any additional keywords are passed into the activity section.

Dizkord.split_message — Method

split_message(text::AbstractString; chunk_limit::UInt=2000,
              extrastyles::Vector{Regex}=Vector{Regex}(),
              forcesplit::Bool = true) -> Vector{String}

Split a message into chunks with at most chunk_limit length, preserving formatting.

The chunk_limit has as default the 2000 character limit of Discord's messages, but can be changed to any nonnegative integer.

Formatting is specified by STYLES) and can be aggregated with the extrastyles argument.

Discord limits messages to 2000, so the code forces split if format breaking cannot be avoided. If desired, however, this behavior can be lifter by setting forcesplit to false.

Examples

```jldoctest; setup=:(using Dizkord) julia> split_message("foo") 1-element Vector{String}: "foo"

julia> split_message(repeat('.', 1995) * "hello, world")[2] "hello, world"

julia> splitmessage("hello, world", chunklimit=10) 2-element Vector{String}: "hello," "world"

julia> splitmessage("hello, _beautiful world", chunklimit=15) ┌ Warning: message was forced-split to fit the desired chunk length limit 15 └ @ Main REPL[66]:28 3-element Vector{String}: "hello," "beautiful wo" "rld_"

julia> splitmessage("hello, _beautiful world", chunklimit=15, forcesplit=false) ┌ Warning: message could not be split into chunks smaller than the length limit 15 └ @ Main REPL[66]:32 2-element Vector{String}: "hello," "beautiful world_"

julia> split_message("hello

", 12) 2-element Vector{String}: "hello ==" "==="

julia> splitmessage("hello ≡≡≡≡≡ ", chunklimit=12, extrastyles = [r" ≡+ "]) 2-element Vector{String}: "hello" "≡≡≡≡≡"

Dizkord.sync_guild_integration — Method

sync_guild_integration(c::Client, guild::Integer, integration::Integer)

Sync an Integration in a Guild.

Dizkord.trigger_typing_indicator — Method

trigger_typing_indicator(c::Client, channel::Integer)

Trigger the typing indicator in a DiscordChannel.

Dizkord.update — Function

update(c::Client, x::T, args...; kwargs...) -> Future{Response}

Update, edit, modify, etc.

Examples

Editing a Message:

update(c, message; content="foo2")

Modifying a Webhook:

update(c, webhook; name="bar2")

Updating a Role:

update(c, role, guild; permissions=8)

Dizkord.update_status — Method

update_status(
    c::Client,
    since::Nullable{Int},
    activity::Nullable{Activity},
    status::Union{PresenceStatus, AbstractString},
    afk::Bool,
) -> Bool

Indicate a presence or status update. A PresenceUpdate event is sent by the gateway in response. More details here.

Dizkord.update_voice_state — Method

update_voice_state(
    c::Client,
    guild::Integer,
    channel::Nullable{Integer},
    mute::Bool,
    deaf::Bool,
) -> Bool

Join, move, or disconnect from a voice channel. A VoiceStateUpdate event is sent by the gateway in response. More details here.

Dizkord.upload_file — Method

upload_file(c::Client, ch::DiscordChannel, path::AbstractString; kwargs...) -> Message

Send a Message with a file Attachment. Any keywords are passed on to create_message.

Dizkord.@deferred_fetch — Macro

@deferred_fetch [functions...] block

Identical to @fetch, but Futures are not fetched until the end of the block. This is more efficient, but only works when there are no data dependencies in the block.

Examples

This will work:

@deferred_fetch begin
    guild_resp = create(c, Guild; name="foo")
    channel_resp = retrieve(c, DiscordChannel, 123)
end

This will not, because the second call is dependent on the first value:

@deferred_fetch begin
    guild_resp = create(c, Guild; name="foo")
    channels_resp = retrieve(c, DiscordChannel, guild_resp.val)
end

Dizkord.@deferred_fetchval — Macro

@deferred_fetchval [functions...] block

Identical to @deferred_fetch, but Futures have fetchval called on them instead of fetch.

Dizkord.@fetch — Macro

@fetch [functions...] block

Wrap all calls to the specified CRUD functions (create, retrieve, update, and delete) with fetch inside a block. If no functions are specified, all CRUD functions are wrapped.

Examples

Wrapping all CRUD functions:

@fetch begin
    guild_resp = create(c, Guild; name="foo")
    guild_resp.ok || error("Request for new guild failed")
    channel_resp = retrieve(c, DiscordChannel, guild_resp.val)
end

Wrapping only calls to retrieve:

@fetch retrieve begin
    resp = retrieve(c, DiscordChannel, 123)
    future = create(c, Message, resp.val; content="foo")  # Behaves normally.
end

Dizkord.@fetchval — Macro

@fetchval [functions...] block

Identical to @fetch, but calls are wrapped with fetchval instead.