Types

Dizkord.Activity — Type

A User activity. More details here.

Fields

name           :: String
type           :: Dizkord.ActivityType
url            :: OptionalNullable{String}
timestamps     :: Optional{Dizkord.ActivityTimestamps}
application_id :: Optional{Snowflake}
details        :: OptionalNullable{String}
state          :: OptionalNullable{String}
emoji          :: OptionalNullable{Dizkord.ActivityEmoji}
party          :: Optional{Dizkord.ActivityParty}
assets         :: Optional{Dizkord.ActivityAssets}
secrets        :: Optional{Dizkord.ActivitySecrets}
instance       :: Optional{Bool}
flags          :: Optional{Int}

Dizkord.ActivityTimestamps — Type

The start and stop times of an Activity. More details here.

Fields

start :: Optional{DateTime}
stop  :: Optional{DateTime}

Dizkord.ActivityParty — Type

The current party of an Activity's player. More details here.

Fields

id   :: Optional{String}
size :: Optional{Vector{Int}}

Dizkord.ActivityAssets — Type

Images and hover text for an Activity. More details here.

Fields

large_image :: Optional{String}
large_text  :: Optional{String}
small_image :: Optional{String}
small_text  :: Optional{String}

Dizkord.ActivitySecrets — Type

Secrets for Rich Presence joining and spectating of an Activity. More details here.

Fields

join     :: Optional{String}
spectate :: Optional{String}
match    :: Optional{String}

Dizkord.ActivityType — Type

An Activity's type. Available values are AT_GAME, AT_STREAMING, AT_LISTENING, AT_WATCHING, and AT_COMPETING. More details here.

Dizkord.ActivityFlags — Type

Flags which indicate what an Activity payload contains. More details here.

Dizkord.ApplicationCommand — Type

An Application Command More details here

Fields

id                  :: OptionalNullable{Snowflake}
type                :: Optional{Dizkord.ApplicationCommandType}
application_id      :: Snowflake
guild_id            :: Optional{Snowflake}
name                :: String
description         :: String
options             :: Optional{Vector{Dizkord.ApplicationCommandOption}}
default_permissions :: Optional{Bool}
version             :: Optional{Snowflake}

Dizkord.ApplicationCommandOption — Type

Application Command Option More details here

Fields

name          :: Optional{String}
type          :: Dizkord.OptionType
value         :: Any
required      :: Optional{Bool}
choices       :: Optional{Vector{Dizkord.ApplicationCommandChoice}}
options       :: Optional{Vector{Dizkord.ApplicationCommandOption}}
channel_types :: Optional{Vector{Dizkord.ChannelTypes}}

Dizkord.ApplicationCommandChoice — Type

Application Command Choice More details here

Fields

name  :: String
value :: Union{Number, String}

Dizkord.Attachment — Type

A Message attachment. More details here.

Fields

id        :: Snowflake
filename  :: String
size      :: Int
url       :: String
proxy_url :: String
height    :: Optional{Int}
width     :: Optional{Int}

Dizkord.AuditLog — Type

An audit log. More details here.

Fields

webhooks          :: Vector{Dizkord.Webhook}
users             :: Vector{Dizkord.User}
audit_log_entries :: Vector{Dizkord.AuditLogEntry}

Dizkord.AuditLogEntry — Type

An entry in an AuditLog. More details here.

Fields

target_id   :: Nullable{Snowflake}
changes     :: Optional{Vector{Dizkord.AuditLogChange}}
user_id     :: Snowflake
id          :: Snowflake
action_type :: Dizkord.ActionType
options     :: Optional{Dizkord.AuditLogOptions}
reason      :: Optional{String}

Dizkord.AuditLogChange — Type

A change item in an AuditLogEntry.

The first type parameter is the type of new_value and old_value. The second is the type of the entity that new_value and old_value belong(ed) to.

More details here.

Fields

new_value :: Optional{T} where T
old_value :: Optional{T} where T
key       :: String
type      :: Type{U} where U

Dizkord.AuditLogOptions — Type

Optional information in an AuditLogEntry. More details here.

Fields

delete_member_days :: Optional{Int}
members_removed    :: Optional{Int}
channel_id         :: Optional{Snowflake}
count              :: Optional{Int}
id                 :: Optional{Snowflake}
type               :: Optional{Dizkord.OverwriteType}
role_name          :: Optional{String}

Dizkord.ActionType — Type

AuditLog action types. More details here.

Dizkord.Ban — Type

A User ban. More details here.

Fields

reason :: Nullable{String}
user   :: Dizkord.User

Dizkord.DiscordChannel — Type

A Discord channel. More details here.

Note: The name Channel is already used, hence the prefix.

Fields

id                    :: Snowflake
type                  :: Dizkord.ChannelType
guild_id              :: Optional{Snowflake}
position              :: Optional{Int}
permission_overwrites :: Optional{Vector{Dizkord.Overwrite}}
name                  :: Optional{String}
topic                 :: OptionalNullable{String}
nsfw                  :: Optional{Bool}
last_message_id       :: OptionalNullable{Snowflake}
bitrate               :: Optional{Int}
user_limit            :: Optional{Int}
rate_limit_per_user   :: Optional{Int}
recipients            :: Optional{Vector{Dizkord.User}}
icon                  :: OptionalNullable{String}
owner_id              :: Optional{Snowflake}
application_id        :: Optional{Snowflake}
parent_id             :: OptionalNullable{Snowflake}
last_pin_timestamp    :: OptionalNullable{DateTime}

Dizkord.ChannelType — Type

A DiscordChannel's type. Available values are CT_GUILD_TEXT, CT_DM, CT_GUILD_VOICE, CT_GROUP_DM, CT_GUILD_CATEGORY, CT_GUILD_NEWS and CT_GUILD_STORE.

Dizkord.Connection — Type

A User connection to an external service (Twitch, YouTube, etc.). More details here.

Fields

id           :: String
name         :: String
type         :: String
revoked      :: Bool
integrations :: Vector{Dizkord.Integration}

Dizkord.Embed — Type

A Message embed. More details here.

Fields

title       :: Optional{String}
type        :: Optional{String}
description :: Optional{String}
url         :: Optional{String}
timestamp   :: Optional{DateTime}
color       :: Optional{Int}
footer      :: Optional{Dizkord.EmbedFooter}
image       :: Optional{Dizkord.EmbedImage}
thumbnail   :: Optional{Dizkord.EmbedThumbnail}
video       :: Optional{Dizkord.EmbedVideo}
provider    :: Optional{Dizkord.EmbedProvider}
author      :: Optional{Dizkord.EmbedAuthor}
fields      :: Optional{Vector{Dizkord.EmbedField}}

Dizkord.EmbedThumbnail — Type

An Embed's thumbnail image information. More details here.

Fields

url       :: Optional{String}
proxy_url :: Optional{String}
height    :: Optional{Int}
width     :: Optional{Int}

Dizkord.EmbedVideo — Type

An Embed's video information. More details here.

Fields

url    :: Optional{String}
height :: Optional{Int}
width  :: Optional{Int}

Dizkord.EmbedImage — Type

An Embed's image information. More details here.

Fields

url       :: Optional{String}
proxy_url :: Optional{String}
height    :: Optional{Int}
width     :: Optional{Int}

Dizkord.EmbedProvider — Type

An Embed's provider information. More details here.

Fields

name :: Optional{String}
url  :: OptionalNullable{String}

Dizkord.EmbedAuthor — Type

An Embed's author information. More details here.

Fields

name           :: Optional{String}
url            :: Optional{String}
icon_url       :: Optional{String}
proxy_icon_url :: Optional{String}

Dizkord.EmbedFooter — Type

An Embed's footer information. More details here.

Fields

text           :: String
icon_url       :: Optional{String}
proxy_icon_url :: Optional{String}

Dizkord.EmbedField — Type

An Embed field. More details here.

Fields

name   :: String
value  :: String
inline :: Optional{Bool}

Dizkord.Emoji — Type

An emoji. More details here.

Fields

id             :: Nullable{Snowflake}
name           :: String
roles          :: Optional{Vector{Snowflake}}
user           :: Optional{Dizkord.User}
require_colons :: Optional{Bool}
managed        :: Optional{Bool}
animated       :: Optional{Bool}

Dizkord.AbstractGuild — Type

A Discord guild (server). Can either be an UnavailableGuild or a Guild.

Dizkord.Guild — Type

A Discord guild (server). More details here.

Fields

id                            :: Snowflake
name                          :: String
icon                          :: Nullable{String}
splash                        :: OptionalNullable{String}
owner                         :: Optional{Bool}
owner_id                      :: Optional{Snowflake}
permissions                   :: Optional{String}
region                        :: Optional{String}
afk_channel_id                :: OptionalNullable{Snowflake}
afk_timeout                   :: Optional{Int}
embed_enabled                 :: Optional{Bool}
embed_channel_id              :: OptionalNullable{Snowflake}
verification_level            :: Optional{Dizkord.VerificationLevel}
default_message_notifications :: Optional{Dizkord.MessageNotificationLevel}
explicit_content_filter       :: Optional{Dizkord.ExplicitContentFilterLevel}
roles                         :: Optional{Vector{Dizkord.Role}}
emojis                        :: Optional{Vector{Dizkord.Emoji}}
features                      :: Optional{Vector{String}}
mfa_level                     :: Optional{Dizkord.MFALevel}
application_id                :: OptionalNullable{Snowflake}
widget_enabled                :: Optional{Bool}
widget_channel_id             :: OptionalNullable{Snowflake}
system_channel_id             :: OptionalNullable{Snowflake}
joined_at                     :: Optional{DateTime}
large                         :: Optional{Bool}
unavailable                   :: Optional{Bool}
member_count                  :: Optional{Int}
max_members                   :: Optional{Int}
voice_states                  :: Optional{Vector{Dizkord.VoiceState}}
members                       :: Optional{Vector{Dizkord.Member}}
channels                      :: Optional{Vector{Dizkord.DiscordChannel}}
presences                     :: Optional{Vector{Dizkord.Presence}}
max_presences                 :: OptionalNullable{Int}
vanity_url_code               :: OptionalNullable{String}
description                   :: OptionalNullable{String}
banner                        :: OptionalNullable{String}

Dizkord.UnavailableGuild — Type

An unavailable Discord guild (server). More details here.

Fields

id          :: Snowflake
unavailable :: Optional{Bool}

Dizkord.VerificationLevel — Type

A Guild's verification level. More details here.

Dizkord.MessageNotificationLevel — Type

A Guild's default message notification level. More details here.

Dizkord.ExplicitContentFilterLevel — Type

A Guild's explicit content filter level. More details here.

Dizkord.MFALevel — Type

A Guild's MFA level. More details here.

Dizkord.GuildEmbed — Type

A Guild embed. More details here.

Fields

enabled    :: Bool
channel_id :: Nullable{Snowflake}

Dizkord.Integration — Type

A Guild integration. More details here.

Fields

id                  :: Snowflake
name                :: String
type                :: String
enabled             :: Bool
syncing             :: Bool
role_id             :: Snowflake
expire_behaviour    :: Int
expire_grace_period :: Int
user                :: Dizkord.User
account             :: Dizkord.IntegrationAccount
synced_at           :: DateTime

Dizkord.IntegrationAccount — Type

An Integration account. More details here.

Fields

id   :: String
name :: String

Dizkord.Interaction — Type

An interaction. More details here.

Fields

id             :: Nullable{Snowflake}
application_id :: Nullable{Snowflake}
type           :: Dizkord.InteractionType
data           :: OptionalNullable{Dizkord.InteractionData}
guild_id       :: Optional{Snowflake}
channel_id     :: Optional{Snowflake}
member         :: Optional{Dizkord.Member}
user           :: Optional{Dizkord.User}
token          :: String
version        :: Optional{Int}
message        :: Optional{Dizkord.Message}

Dizkord.InteractionData — Type

Data for an interaction. More details here.

Fields

id             :: Nullable{Snowflake}
name           :: String
type           :: Dizkord.ApplicationCommandType
resolved       :: Optional{Dizkord.ResolvedData}
options        :: Optional{Vector{Dizkord.ApplicationCommandOption}}
custom_id      :: OptionalNullable{String}
component_type :: OptionalNullable{Dizkord.ComponentType}
values         :: Optional{Vector{String}}
target_id      :: Optional{Snowflake}

Dizkord.Invite — Type

An invite to a Guild. More details here.

Fields

code                       :: String
guild                      :: Optional{Dizkord.Guild}
channel                    :: Dizkord.DiscordChannel
approximate_presence_cound :: Optional{Int}
approximate_member_count   :: Optional{Int}

Dizkord.InviteMetadata — Type

Metadata for an Invite. More details here.

Fields

inviter    :: Dizkord.User
uses       :: Int
max_uses   :: Int
max_age    :: Int
temporary  :: Bool
created_at :: DateTime
revoked    :: Bool

Dizkord.Member — Type

A Guild member. More details here.

Fields

user          :: Optional{Dizkord.User}
nick          :: OptionalNullable{String}
roles         :: Vector{Snowflake}
joined_at     :: DateTime
premium_since :: OptionalNullable{DateTime}
deaf          :: Bool
mute          :: Bool

Dizkord.Message — Type

A message sent to a DiscordChannel. More details here.

Fields

id               :: Snowflake
channel_id       :: Snowflake
guild_id         :: Optional{Snowflake}
author           :: Optional{Dizkord.User}
member           :: Optional{Dizkord.Member}
content          :: Optional{String}
timestamp        :: Optional{DateTime}
edited_timestamp :: OptionalNullable{DateTime}
tts              :: Optional{Bool}
mention_everyone :: Optional{Bool}
mentions         :: Optional{Vector{Dizkord.User}}
mention_roles    :: Optional{Vector{Snowflake}}
attachments      :: Optional{Vector{Dizkord.Attachment}}
embeds           :: Optional{Vector{Dizkord.Embed}}
reactions        :: Optional{Vector{Dizkord.Reaction}}
nonce            :: OptionalNullable{Snowflake}
pinned           :: Optional{Bool}
webhook_id       :: Optional{Snowflake}
type             :: Optional{Dizkord.MessageType}
activity         :: Optional{Dizkord.MessageActivity}
application      :: Optional{Dizkord.MessageApplication}

Dizkord.MessageActivity — Type

A Message activity. More details here.

Fields

type     :: Dizkord.MessageActivityType
party_id :: Optional{String}

Dizkord.MessageApplication — Type

A Rich Presence Message's application information. More details here.

Fields

id          :: Snowflake
cover_image :: Optional{String}
description :: String
icon        :: String
name        :: String

Dizkord.MessageType — Type

A Message's type. More details here.

Dizkord.MessageActivityType — Type

A Message's activity type. More details here.

Dizkord.Overwrite — Type

A permission overwrite. More details here.

Fields

id    :: Snowflake
type  :: Dizkord.OverwriteType
allow :: String
deny  :: String

Dizkord.OverwriteType — Type

An Overwrite's type. More details here.

Dizkord.Presence — Type

A User's presence. More details here.

Fields

user       :: Dizkord.User
roles      :: Optional{Vector{Snowflake}}
game       :: Nullable{Dizkord.Activity}
guild_id   :: Optional{Snowflake}
status     :: Dizkord.PresenceStatus
activities :: Vector{Dizkord.Activity}

Dizkord.PresenceStatus — Type

A User's status sent in a Presence. Available values are PS_IDLE, PS_DND, PS_ONLINE, and PS_OFFLINE. More details here.

Dizkord.Reaction — Type

A Message reaction. More details here.

Fields

count :: Int
me    :: Bool
emoji :: Dizkord.Emoji

Dizkord.Role — Type

A User role. More details here.

Fields

id          :: Snowflake
name        :: String
color       :: Optional{Int}
hoist       :: Optional{Bool}
position    :: Optional{Int}
permissions :: Optional{String}
managed     :: Optional{Bool}
mentionable :: Optional{Bool}

Dizkord.User — Type

A Discord user. More details here.

Fields

id            :: Snowflake
username      :: Optional{String}
discriminator :: Optional{String}
avatar        :: OptionalNullable{String}
bot           :: Optional{Bool}
mfa_enabled   :: Optional{Bool}
locale        :: Optional{String}
verified      :: Optional{Bool}
email         :: OptionalNullable{String}

Dizkord.VoiceRegion — Type

A region for a Guild's voice server. More details here.

Fields

id         :: String
name       :: String
vip        :: Bool
optimal    :: Bool
deprecated :: Bool
custom     :: Bool

Dizkord.VoiceState — Type

A User's voice connection status. More details here.

Fields

guild_id   :: Optional{Snowflake}
channel_id :: Nullable{Snowflake}
user_id    :: Snowflake
member     :: Optional{Dizkord.Member}
session_id :: String
deaf       :: Bool
mute       :: Bool
self_deaf  :: Bool
self_mute  :: Bool
suppress   :: Bool

Dizkord.Webhook — Type

A Webhook. More details here.

Fields

id         :: Snowflake
guild_id   :: Optional{Snowflake}
channel_id :: Snowflake
user       :: Optional{Dizkord.User}
name       :: Nullable{String}
avatar     :: Nullable{String}
token      :: Optional{String}