REST API

Response

Dizkord.Response — Type

A wrapper around a response from the REST API. Every function which wraps a Discord REST API endpoint returns a Future which will contain a value of this type. To retrieve the Response from the Future, use fetch. See also: fetchval.

Fields

val::Nullable{T}: The object contained in the HTTP response. For example, for a call to get_channel_message, this value will be a Message.

ok::Bool: The state of the request. If true, then it is safe to access val.

http_response::Nullable{HTTP.Messages.Response}: The underlying HTTP response, if a request was made.

exception::Nullable{Exception}: The caught exception, if one is thrown.

Examples

Multiple API calls which immediately return Futures and can be awaited:

futures = map(i -> create_message(c, channel_id; content=string(i)), 1:10);
other_work_here()
resps = fetch.(futures)

Skipping error checks and returning the value directly:

guild = fetchval(create_guild(c; name="foo"))

Dizkord.fetchval — Function

fetchval(f::Future{Response{T}}) -> Nullable{T}

Shortcut for fetch(f).val: Fetch a Response and return its value. Note that there are no guarantees about the response's success and the value being returned, and it discards context that can be useful for debugging, such as HTTP responses and caught exceptions.

CRUD API

On top of functions for accessing individual endpoints such as get_channel_messages, Dizkord.jl also offers a unified API with just four functions. Named after the CRUD model, they cover most of the Discord REST API and allow you to write concise, expressive code, and forget about the subtleties of endpoint naming. The argument ordering convention is roughly as follows:

A Client, always.

For cases when we don't yet have the entity to be manipulated (usually create and retrieve), the entity's type. If we do have the entity (update and delete), the entity itself.

The remaining positional arguments supply whatever context is needed to specify the entity. For example, sending a message requires a DiscordChannel parameter.

Keyword arguments follow (usually for create and update).

Dizkord.create — Function

create(c::Client, ::Type{T}, args...; kwargs...) -> Future{Response}

Create, add, send, etc.

Examples

Sending a Message:

create(c, Message, channel; content="foo")

Creating a new DiscordChannel:

create(c, DiscordChannel, guild; name="bar")

Banning a Member:

create(c, Ban, guild, member; reason="baz")

Dizkord.retrieve — Function

retrieve(c::Client, ::Type{T}, args...; kwargs...) -> Future{Response}

Retrieve, get, list, etc.

Examples

Getting the Client's User:

retrieve(c, User)

Getting a Guild's DiscordChannels:

retrieve(c, DiscordChannel, guild)

Getting an Invite to a Guild by code:

retrieve(c, Invite, "abcdef")

Dizkord.update — Function

update(c::Client, x::T, args...; kwargs...) -> Future{Response}

Update, edit, modify, etc.

Examples

Editing a Message:

update(c, message; content="foo2")

Modifying a Webhook:

update(c, webhook; name="bar2")

Updating a Role:

update(c, role, guild; permissions=8)

Dizkord.delete — Function

delete(c::Client, x::T, args...) -> Future{Response}

Delete, remove, discard, etc.

Examples

Kicking a Member:

delete(c, member)

Unbanning a Member:

delete(c, ban, guild)

Deleting all Reactions from a Message (note: this is the only update/delete method which takes a type parameter):

delete(c, Reaction, message)

The full list of types available to be manipulated is:

AuditLog

ApplicationCommand

Ban

DiscordChannel

Emoji

GuildEmbed

Guild

Integration

Invite

Member

Message

Overwrite

Reaction

Role

User

VoiceRegion

Webhook

Endpoints

Functions which wrap REST API endpoints are named and sorted according to the Discord API documentation. When a function accepts keyword arguments, the docstring will include a link to the Discord documentation which indicates the expected keys and values. Remember that the return types annotated below are not the actual return types, but the types of Response that the returned Futures will yield.

Audit Log

Dizkord.get_guild_audit_log — Function

get_guild_audit_log(c::Client, guild::Integer; kwargs...) -> AuditLog

Get a Guild's AuditLog. More details here.

Channel

Dizkord.get_channel — Function

get_channel(c::Client, channel::Integer) -> DiscordChannel

Get a DiscordChannel.

Dizkord.modify_channel — Function

modify_channel(c::Client, channel::Integer; kwargs...) -> DiscordChannel

Modify a DiscordChannel. More details here.

Dizkord.delete_channel — Function

delete_channel(c::Client, channel::Integer) -> DiscordChannel

Delete a DiscordChannel.

Dizkord.get_channel_messages — Function

get_channel_messages(c::Client, channel::Integer; kwargs...) -> Vector{Message}

Get a list of Messages from a DiscordChannel. More details here.

Dizkord.get_channel_message — Function

get_channel_message(c::Client, channel::Integer, message::Integer) -> Message

Get a Message from a DiscordChannel.

Dizkord.create_message — Function

create_message(c::Client, channel::Integer; kwargs...) -> Message

Send a Message to a DiscordChannel. More details here.

Dizkord.create_reaction — Function

create_reaction(c::Client, channel::Integer, message::Integer, emoji::StringOrChar)

React to a Message. If emoji is a custom Emoji, it should be formatted "name:id".

Dizkord.delete_own_reaction — Function

delete_own_reaction(c::Client, channel::Integer, message::Integer, emoji::StringOrChar)

Delete the Client user's reaction to a Message.

Dizkord.delete_user_reaction — Function

delete_user_reaction(
    c::Client,
    channel::Integer,
    message::Integer,
    emoji::StringOrChar,
    user::Integer,
)

Delete a User's reaction to a Message.

Dizkord.get_reactions — Function

get_reactions(
    c::Client,
    channel::Integer,
    message::Integer,
    emoji::StringOrChar,
) -> Vector{User}

Get the Users who reacted to a Message with an Emoji.

Dizkord.delete_all_reactions — Function

delete_all_reactions(c::Client, channel::Integer, message::Integer)

Delete all reactions from a Message.

Dizkord.edit_message — Function

edit_message(c::Client, channel::Integer, message::Integer; kwargs...) -> Message

Edit a Message. More details here.

Dizkord.delete_message — Function

delete_message(c::Client, channel::Integer, message::Integer)

Delete a Message.

Dizkord.bulk_delete_messages — Function

bulk_delete_messages(c::Client, channel::Integer; kwargs...)

Delete multiple Messages. More details here.

Dizkord.edit_channel_permissions — Function

edit_channel_permissions(
    c::Client,
    channel::Integer,
    overwrite::Integer;
    kwargs...,
)

Edit permissions for a DiscordChannel. More details here.

Dizkord.get_channel_invites — Function

get_channel_invites(c::Client, channel::Integer) -> Vector{Invite}

Get the Invites for a DiscordChannel.

Dizkord.create_channel_invite — Function

create_channel_invite(c::Client, channel::Integer; kwargs...) -> Invite

Create an Invite to a DiscordChannel. More details here.

Dizkord.delete_channel_permission — Function

delete_channel_permission(c::Client, channel::Integer, overwrite::Integer)

Delete an Overwrite from a DiscordChannel.

Dizkord.trigger_typing_indicator — Function

trigger_typing_indicator(c::Client, channel::Integer)

Trigger the typing indicator in a DiscordChannel.

Dizkord.get_pinned_messages — Function

get_pinned_messages(c::Client, channel::Integer) -> Vector{Message}

Get the pinned Messages in a DiscordChannel.

Dizkord.add_pinned_channel_message — Function

add_pinned_channel_message(c::Client, channel::Integer, message::Integer)

Pin a Message in a DiscordChannel.

Dizkord.delete_pinned_channel_message — Function

delete_pinned_channel_message(c::Client, channel::Integer, message::Integer)

Unpin a Message from a DiscordChannel.

Emoji

Dizkord.list_guild_emojis — Function

list_guild_emojis(c::Client, guild::Integer) -> Vector{Emoji}

Get the Emojis in a Guild.

Dizkord.get_guild_emoji — Function

get_guild_emoji(c::Client, guild::Integer, emoji::Integer) -> Emoji

Get an Emoji in a Guild.

Dizkord.create_guild_emoji — Function

create_guild_emoji(c::Client, guild::Integer; kwargs...) -> Emoji

Create an Emoji in a Guild. More details here.

Dizkord.modify_guild_emoji — Function

modify_guild_emoji(c::Client, guild::Integer, emoji::Integer; kwargs...) -> Emoji

Edit an Emoji in a Guild. More details here.

Dizkord.delete_guild_emoji — Function

delete_guild_emoji(c::Client, guild::Integer, emoji::Integer)

Delete an Emoji from a Guild.

Guild

Dizkord.create_guild — Function

create_guild(c::Client; kwargs...) -> Guild

Create a Guild. More details here.

Dizkord.get_guild — Function

get_guild(c::Client, guild::Integer) -> Guild

Get a Guild.

Dizkord.modify_guild — Function

modify_guild(c::Client, guild::Integer; kwargs...) -> Guild

Edit a Guild. More details here.

Dizkord.delete_guild — Function

delete_guild(c::Client, guild::Integer)

Delete a Guild.

Dizkord.get_guild_channels — Function

get_guild_channels(c::Client, guild::Integer) -> Vector{DiscordChannel}

Get the DiscordChannels in a Guild.

Dizkord.create_guild_channel — Function

create_guild_channel(c::Client, guild::Integer; kwargs...) -> DiscordChannel

Create a DiscordChannel in a Guild. More details here.

Dizkord.modify_guild_channel_positions — Function

modify_guild_channel_positions(c::Client, guild::Integer, positions...)

Modify the positions of DiscordChannels in a Guild. More details here.

Dizkord.get_guild_member — Function

get_guild_member(c::Client, guild::Integer, user::Integer) -> Member

Get a Member in a Guild.

Dizkord.list_guild_members — Function

list_guild_members(c::Client, guild::Integer; kwargs...) -> Vector{Member}

Get a list of Members in a Guild. More details here.

Dizkord.add_guild_member — Function

add_guild_member(c::Client; kwargs...) -> Member

Add a User to a Guild. More details here.

Dizkord.modify_guild_member — Function

modify_guild__member(c::Client, guild::Integer, user::Integer; kwargs...)

Modify a Member in a Guild. More details here.

Dizkord.modify_current_user_nick — Function

modify_current_user_nick(c::Client, guild::Intger; kwargs...) -> String

Modify the Client user's nickname in a Guild. More details here.

Dizkord.add_guild_member_role — Function

add_guild_member_role(c::Client, guild::Integer, user::Integer, role::Integer)

Add a Role to a Member.

Dizkord.remove_guild_member_role — Function

remove_guild_member_role(c::Client, guild::Integer, user::Integer, role::Integer)

Remove a Role from a Member.

Dizkord.remove_guild_member — Function

remove_guild_member(c::Client, guild::Integer, user::Integer)

Kick a Member from a Guild.

Dizkord.get_guild_bans — Function

get_guild_bans(c::Client, guild::Integer) -> Vector{Ban}

Get a list of Bans in a Guild.

Dizkord.get_guild_ban — Function

get_ban(c::Client, guild::Integer,  user::Integer) -> Ban

Get a Ban in a Guild.

Dizkord.create_guild_ban — Function

create_guild_ban(c::Client, guild::Integer, user::Integer; kwargs...)

Ban a Member from a Guild. More details here.

Dizkord.remove_guild_ban — Function

remove_guild_ban(c::Client, guild::Integer, user::Integer)

Unban a User from a Guild.

Dizkord.get_guild_roles — Function

get_guild_roles(c::Client, guild::Integer) -> Vector{Role}

Get a Guild's Roles.

Dizkord.create_guild_role — Function

create_guild_role(c::Client, guild::Integer; kwargs) -> Role

Create a Role in a Guild. More details here.

Dizkord.modify_guild_role_positions — Function

modify_guild_role_positions(c::Client, guild::Integer, positions...) -> Vector{Role}

Modify the positions of Roles in a Guild. More details here.

Dizkord.modify_guild_role — Function

modify_guild_role(c::Client, guild::Integer, role::Integer; kwargs) -> Role

Modify a Role in a Guild. More details here.

Dizkord.delete_guild_role — Function

delete_guild_role(c::Client, guild::Integer, role::Integer)

Delete a Role from a Guild.

Dizkord.get_guild_prune_count — Function

get_guild_prune_count(c::Client, guild::Integer; kwargs...) -> Dict

Get the number of Members that would be removed from a Guild in a prune. More details here.

Dizkord.begin_guild_prune — Function

begin_guild_prune(c::Client, guild::Integer; kwargs...) -> Dict

Begin pruning Members from a Guild. More details here.

Dizkord.get_guild_voice_regions — Function

get_guild_voice_regions(c::Client, guild::Integer) -> Vector{VoiceRegion}

Get a list of VoiceRegions for the Guild.

Dizkord.get_guild_invites — Function

get_guild_invites(c::Client, guild::Integer) -> Vector{Invite}

Get a list of Invites to a Guild.

Dizkord.get_guild_integrations — Function

get_guild_integrations(c::Client, guild::Integer) -> Vector{Integration}

Get a list of Integrations for a Guild.

Dizkord.create_guild_integration — Function

create_guild_integration(c::Client, guild::Integer; kwargs...)

Create/attach an Integration to a Guild. More details here.

Dizkord.modify_guild_integration — Function

modify_guild_integration(c::Client, guild::Integer, integration::Integer; kwargs...)

Modify an Integration in a Guild. More details here.

Dizkord.delete_guild_integration — Function

delete_guild_integration(c::Client, guild::Integer, integration::Integer)

Delete an Integration from a Guild.

Dizkord.sync_guild_integration — Function

sync_guild_integration(c::Client, guild::Integer, integration::Integer)

Sync an Integration in a Guild.

Dizkord.get_guild_embed — Function

get_guild_embed(c::Client, guild::Integer) -> GuildEmbed

Get a Guild's GuildEmbed.

Dizkord.modify_guild_embed — Function

modify_guild_embed(c::Client, guild::Integer; kwargs...) -> GuildEmbed

Modify a Guild's GuildEmbed. More details here.

Dizkord.get_vanity_url — Function

get_vanity_url(c::Client, guild::Integer) -> Invite

Get a Guild's vanity URL, if it supports that feature.

Dizkord.get_guild_widget_image — Function

get_guild_widget_image(c::Client, guild::Integer; kwargs...) -> Vector{UInt8}

Get a Guild's widget image in PNG format. More details here.

Invite

Dizkord.get_invite — Function

get_invite(c::Client, invite::AbstractString; kwargs...} -> Invite

Get an Invite to a Guild. More details here.

Dizkord.delete_invite — Function

delete_invite(c::Client, invite::AbstractString) -> Invite

Delete an Invite to a Guild.

User

Dizkord.get_current_user — Function

get_current_user(c::Client) -> User

Get the Client User.

Dizkord.get_user — Function

get_user(c::Client, user::Integer) -> User

Get a User.

Dizkord.modify_current_user — Function

modify_current_user(c::Client; kwargs...) -> User

Modify the Client User. More details here.

Dizkord.get_current_user_guilds — Function

get_user_guilds(c::Client; kwargs...) -> Vector{Guild}

Get a list of Guilds the Client User is a member of. More details here.

Dizkord.leave_guild — Function

leave_guild(c::Client, guild::Integer)

Leave a Guild.

Dizkord.create_dm — Function

create_dm(c::Client; kwargs...) -> DiscordChannel

Create a DM DiscordChannel. More details here.

Voice

Dizkord.list_voice_regions — Function

list_voice_regions(c::Client) -> Vector{VoiceRegion}

Get a list of the VoiceRegions that can be used when creating Guilds.

Webhook

Dizkord.create_webhook — Function

create_webhook(c::Client, channel::Integer; kwargs...) -> Webhook

Create a Webhook in a DiscordChannel. More details here.

Dizkord.get_channel_webhooks — Function

get_channel_webhooks(c::Client, channel::Integer) -> Vector{Webhook}

Get a list of Webhooks in a DiscordChannel.

Dizkord.get_guild_webhooks — Function

get_guild_webhooks(c::Client, guild::Integer) -> Vector{Webhook}

Get a list of Webhooks in a Guild.

Dizkord.get_webhook — Function

get_webhook(c::Client, webhook::Integer) -> Webhook

Get a Webhook.

Dizkord.get_webhook_with_token — Function

get_webhook_with_token(c::Client, webhook::Integer, token::AbstractString) -> Webhook

Get a Webhook with a token.

Dizkord.modify_webhook — Function

modify_webhook(c::Client, webhook::Integer; kwargs...) -> Webhook

Modify a Webhook. More details here.

Dizkord.modify_webhook_with_token — Function

modify_webhook_with_token(
    c::Client,
    webhook::Integer,
    token::AbstractString;
    kwargs...,
) -> Webhook

Modify a Webhook with a token. More details here.

Dizkord.delete_webhook — Function

delete_webhook(c::Client, webhook::Integer)

Delete a Webhook.

Dizkord.delete_webhook_with_token — Function

delete_webhook_with_token(c::Client, webhook::Integer, token::AbstractString)

Delete a Webhook with a token.

Dizkord.execute_webhook — Function

execute_webhook(
    c::Client,
    webhook::Integer,
    token::AbstractString;
    wait::Bool=false,
    kwargs...,
) -> Message

Execute a Webhook. If wait is not set, no Message is returned. More details here.

Dizkord.execute_slack_compatible_webhook — Function

execute_slack_compatible_webhook(
    c::Client,
    webhook::Integer,
    token::AbstractString;
    wait::Bool=true,
    kwargs...,
)

Execute a Slack Webhook. More details here.

Dizkord.execute_github_compatible_webhook — Function

execute_github_compatible_webhook(
    c::Client,
    webhook::Integer,
    token::AbstractString;
    wait::Bool=true,
    kwargs...,
)

Execute a Github Webhook. More details here.