Types

D1zk0rd.Message — Type

A message sent to a DiscordChannel. More details here.

Fields

id               :: Snowflake
channel_id       :: Snowflake
guild_id         :: Optional{Snowflake}
author           :: Optional{D1zk0rd.User}
member           :: Optional{D1zk0rd.Member}
content          :: Optional{String}
timestamp        :: Optional{DateTime}
edited_timestamp :: OptionalNullable{DateTime}
tts              :: Optional{Bool}
mention_everyone :: Optional{Bool}
mentions         :: Optional{Vector{D1zk0rd.User}}
mention_roles    :: Optional{Vector{Snowflake}}
attachments      :: Optional{Vector{D1zk0rd.Attachment}}
embeds           :: Optional{Vector{D1zk0rd.Embed}}
reactions        :: Optional{Vector{D1zk0rd.Reaction}}
nonce            :: OptionalNullable{Snowflake}
pinned           :: Optional{Bool}
webhook_id       :: Optional{Snowflake}
type             :: Optional{Int}
activity         :: Optional{D1zk0rd.MessageActivity}
application      :: Optional{D1zk0rd.MessageApplication}

D1zk0rd.AbstractGuild — Type

A Discord guild (server). Can either be an UnavailableGuild or a Guild.

D1zk0rd.Guild — Type

A Discord guild (server). More details here.

Fields

id                            :: Snowflake
name                          :: String
icon                          :: Nullable{String}
splash                        :: OptionalNullable{String}
owner                         :: Optional{Bool}
owner_id                      :: Optional{Snowflake}
permissions                   :: Optional{String}
region                        :: Optional{String}
afk_channel_id                :: OptionalNullable{Snowflake}
afk_timeout                   :: Optional{Int}
embed_enabled                 :: Optional{Bool}
embed_channel_id              :: OptionalNullable{Snowflake}
verification_level            :: Optional{Int}
default_message_notifications :: Optional{Int}
explicit_content_filter       :: Optional{Int}
roles                         :: Optional{Vector{D1zk0rd.Role}}
emojis                        :: Optional{Vector{D1zk0rd.Emoji}}
features                      :: Optional{Vector{String}}
mfa_level                     :: Optional{Int}
application_id                :: OptionalNullable{Snowflake}
widget_enabled                :: Optional{Bool}
widget_channel_id             :: OptionalNullable{Snowflake}
system_channel_id             :: OptionalNullable{Snowflake}
joined_at                     :: Optional{DateTime}
large                         :: Optional{Bool}
unavailable                   :: Optional{Bool}
member_count                  :: Optional{Int}
max_members                   :: Optional{Int}
voice_states                  :: Optional{Vector{D1zk0rd.VoiceState}}
members                       :: Optional{Vector{D1zk0rd.Member}}
channels                      :: Optional{Vector{D1zk0rd.DiscordChannel}}
presences                     :: Optional{Vector{D1zk0rd.Presence}}
max_presences                 :: OptionalNullable{Int}
vanity_url_code               :: OptionalNullable{String}
description                   :: OptionalNullable{String}
banner                        :: OptionalNullable{String}

D1zk0rd.Activity — Type

A User activity. More details here.

Fields

name           :: String
type           :: Int
url            :: OptionalNullable{String}
timestamps     :: Optional{D1zk0rd.ActivityTimestamps}
application_id :: Optional{Snowflake}
details        :: OptionalNullable{String}
state          :: OptionalNullable{String}
emoji          :: OptionalNullable{D1zk0rd.ActivityEmoji}
party          :: Optional{D1zk0rd.ActivityParty}
assets         :: Optional{D1zk0rd.ActivityAssets}
secrets        :: Optional{D1zk0rd.ActivitySecrets}
instance       :: Optional{Bool}
flags          :: Optional{Int}

D1zk0rd.ActivityTimestamps — Type

The start and stop times of an Activity. More details here.

Fields

start :: Optional{DateTime}
stop  :: Optional{DateTime}

D1zk0rd.ActivityParty — Type

The current party of an Activity's player. More details here.

Fields

id   :: Optional{String}
size :: Optional{Vector{Int}}

D1zk0rd.ActivityAssets — Type

Images and hover text for an Activity. More details here.

Fields

large_image :: Optional{String}
large_text  :: Optional{String}
small_image :: Optional{String}
small_text  :: Optional{String}

D1zk0rd.ActivitySecrets — Type

Secrets for Rich Presence joining and spectating of an Activity. More details here.

Fields

join     :: Optional{String}
spectate :: Optional{String}
match    :: Optional{String}

D1zk0rd.ActivityType — Module

An Activity's type. Available values are GAME, STREAMING, LISTENING, WATCHING, and COMPETING. More details here.

D1zk0rd.ActivityFlags — Module

Flags which indicate what an Activity payload contains. More details here.

D1zk0rd.ApplicationCommand — Type

An Application Command. More details here.

Fields

id                  :: OptionalNullable{Snowflake}
type                :: Optional{Int}
application_id      :: Snowflake
guild_id            :: Optional{Snowflake}
name                :: String
description         :: String
options             :: Optional{Vector{D1zk0rd.ApplicationCommandOption}}
default_permissions :: Optional{Bool}
version             :: Optional{Snowflake}

D1zk0rd.ApplicationCommandOption — Type

Application Command Option. More details here.

Fields

value         :: Any
type          :: Optional{Int}
name          :: Optional{String}
description   :: Optional{String}
required      :: Optional{Bool}
min_value     :: Optional{Number}
max_value     :: Optional{Number}
autocomplete  :: Optional{Bool}
choices       :: Optional{Vector{D1zk0rd.ApplicationCommandChoice}}
options       :: Optional{Vector{D1zk0rd.ApplicationCommandOption}}
channel_types :: Optional{Vector{Int}}
focused       :: Optional{Bool}

D1zk0rd.ApplicationCommandChoice — Type

Application Command Choice. More details here.

Fields

name  :: String
value :: Union{Number, String}

D1zk0rd.Attachment — Type

A Message attachment. More details here.

Fields

id        :: Snowflake
filename  :: String
size      :: Int
url       :: String
proxy_url :: String
height    :: Optional{Int}
width     :: Optional{Int}

D1zk0rd.AuditLog — Type

An audit log. More details here.

Fields

webhooks          :: Vector{D1zk0rd.Webhook}
users             :: Vector{D1zk0rd.User}
audit_log_entries :: Vector{D1zk0rd.AuditLogEntry}

D1zk0rd.AuditLogEntry — Type

An entry in an AuditLog. More details here.

Fields

target_id   :: Nullable{Snowflake}
changes     :: Optional{Vector{D1zk0rd.AuditLogChange}}
user_id     :: Snowflake
id          :: Snowflake
action_type :: Int
options     :: Optional{D1zk0rd.AuditLogOptions}
reason      :: Optional{String}

D1zk0rd.AuditLogChange — Type

A change item in an AuditLogEntry.

The first type parameter is the type of new_value and old_value. The second is the type of the entity that new_value and old_value belong(ed) to.

More details here.

Fields

new_value :: Optional{T} where T
old_value :: Optional{T} where T
key       :: String
type      :: Type{U} where U

D1zk0rd.AuditLogOptions — Type

Optional information in an AuditLogEntry. More details here.

Fields

delete_member_days :: Optional{Int}
members_removed    :: Optional{Int}
channel_id         :: Optional{Snowflake}
count              :: Optional{Int}
id                 :: Optional{Snowflake}
type               :: Optional{Int}
role_name          :: Optional{String}

D1zk0rd.ActionType — Module

AuditLog action types. More details here.

D1zk0rd.Ban — Type

A User ban. More details here.

Fields

reason :: Nullable{String}
user   :: D1zk0rd.User

D1zk0rd.DiscordChannel — Type

A Discord channel. More details here.

Note: The name Channel is already used, hence the prefix.

Fields

id                    :: Snowflake
type                  :: Int
guild_id              :: Optional{Snowflake}
position              :: Optional{Int}
permission_overwrites :: Optional{Vector{D1zk0rd.Overwrite}}
name                  :: Optional{String}
topic                 :: OptionalNullable{String}
nsfw                  :: Optional{Bool}
last_message_id       :: OptionalNullable{Snowflake}
bitrate               :: Optional{Int}
user_limit            :: Optional{Int}
rate_limit_per_user   :: Optional{Int}
recipients            :: Optional{Vector{D1zk0rd.User}}
icon                  :: OptionalNullable{String}
owner_id              :: Optional{Snowflake}
application_id        :: Optional{Snowflake}
parent_id             :: OptionalNullable{Snowflake}
last_pin_timestamp    :: OptionalNullable{DateTime}

D1zk0rd.Connection — Type

A User connection to an external service (Twitch, YouTube, etc.). More details here.

Fields

id           :: String
name         :: String
type         :: String
revoked      :: Bool
integrations :: Vector{D1zk0rd.Integration}

D1zk0rd.Component — Type

An interactable component. More details here.

Fields

type        :: Int
custom_id   :: Optional{String}
disabled    :: Optional{Bool}
style       :: Optional{Int}
label       :: Optional{String}
emoji       :: Optional{D1zk0rd.Emoji}
url         :: Optional{String}
options     :: Optional{Vector{D1zk0rd.SelectOption}}
placeholder :: Optional{String}
min_values  :: Optional{Int}
max_values  :: Optional{Int}
components  :: Optional{Vector{D1zk0rd.Component}}

D1zk0rd.Embed — Type

A Message embed. More details here.

Fields

title       :: Optional{String}
type        :: Optional{String}
description :: Optional{String}
url         :: Optional{String}
timestamp   :: Optional{DateTime}
color       :: Optional{Int}
footer      :: Optional{D1zk0rd.EmbedFooter}
image       :: Optional{D1zk0rd.EmbedImage}
thumbnail   :: Optional{D1zk0rd.EmbedThumbnail}
video       :: Optional{D1zk0rd.EmbedVideo}
provider    :: Optional{D1zk0rd.EmbedProvider}
author      :: Optional{D1zk0rd.EmbedAuthor}
fields      :: Optional{Vector{D1zk0rd.EmbedField}}

D1zk0rd.EmbedThumbnail — Type

An Embed's thumbnail image information. More details here.

Fields

url       :: Optional{String}
proxy_url :: Optional{String}
height    :: Optional{Int}
width     :: Optional{Int}

D1zk0rd.EmbedVideo — Type

An Embed's video information. More details here.

Fields

url    :: Optional{String}
height :: Optional{Int}
width  :: Optional{Int}

D1zk0rd.EmbedImage — Type

An Embed's image information. More details here.

Fields

url       :: Optional{String}
proxy_url :: Optional{String}
height    :: Optional{Int}
width     :: Optional{Int}

D1zk0rd.EmbedProvider — Type

An Embed's provider information. More details here.

Fields

name :: Optional{String}
url  :: OptionalNullable{String}

D1zk0rd.EmbedAuthor — Type

An Embed's author information. More details here.

Fields

name           :: Optional{String}
url            :: Optional{String}
icon_url       :: Optional{String}
proxy_icon_url :: Optional{String}

D1zk0rd.EmbedFooter — Type

An Embed's footer information. More details here.

Fields

text           :: String
icon_url       :: Optional{String}
proxy_icon_url :: Optional{String}

D1zk0rd.EmbedField — Type

An Embed field. More details here.

Fields

name   :: String
value  :: String
inline :: Optional{Bool}

D1zk0rd.Emoji — Type

An emoji. More details here.

Fields

id             :: Nullable{Snowflake}
name           :: String
roles          :: Optional{Vector{Snowflake}}
user           :: Optional{D1zk0rd.User}
require_colons :: Optional{Bool}
managed        :: Optional{Bool}
animated       :: Optional{Bool}

D1zk0rd.UnavailableGuild — Type

An unavailable Discord guild (server). More details here.

Fields

id          :: Snowflake
unavailable :: Optional{Bool}

D1zk0rd.VerificationLevel — Module

A Guild's verification level. More details here.

D1zk0rd.MessageNotificationLevel — Module

A Guild's default message notification level. More details here.

D1zk0rd.ExplicitContentFilterLevel — Module

A Guild's explicit content filter level. More details here.

D1zk0rd.MFALevel — Module

A Guild's MFA level. More details here.

D1zk0rd.GuildEmbed — Type

A Guild embed. More details here.

Fields

enabled    :: Bool
channel_id :: Nullable{Snowflake}

D1zk0rd.Integration — Type

A Guild integration. More details here.

Fields

id                  :: Snowflake
name                :: String
type                :: String
enabled             :: Bool
syncing             :: Bool
role_id             :: Snowflake
expire_behaviour    :: Int
expire_grace_period :: Int
user                :: D1zk0rd.User
account             :: D1zk0rd.IntegrationAccount
synced_at           :: DateTime

D1zk0rd.IntegrationAccount — Type

An Integration account. More details here.

Fields

id   :: String
name :: String

D1zk0rd.Interaction — Type

An interaction. More details here.

Fields

id             :: Nullable{Snowflake}
application_id :: Nullable{Snowflake}
type           :: Int
data           :: OptionalNullable{D1zk0rd.InteractionData}
guild_id       :: Optional{Snowflake}
channel_id     :: Optional{Snowflake}
member         :: Optional{D1zk0rd.Member}
user           :: Optional{D1zk0rd.User}
token          :: String
version        :: Optional{Int}
message        :: Optional{D1zk0rd.Message}

D1zk0rd.InteractionData — Type

Data for an interaction. More details here.

Fields

id             :: OptionalNullable{Snowflake}
name           :: OptionalNullable{String}
type           :: OptionalNullable{Int}
resolved       :: Optional{D1zk0rd.ResolvedData}
options        :: Optional{Vector{D1zk0rd.ApplicationCommandOption}}
custom_id      :: OptionalNullable{String}
component_type :: OptionalNullable{Int}
values         :: Optional{Vector{String}}
target_id      :: Optional{Snowflake}

D1zk0rd.Invite — Type

An invite to a Guild. More details here.

Fields

code                       :: String
guild                      :: Optional{D1zk0rd.Guild}
channel                    :: D1zk0rd.DiscordChannel
approximate_presence_cound :: Optional{Int}
approximate_member_count   :: Optional{Int}

D1zk0rd.InviteMetadata — Type

Metadata for an Invite. More details here.

Fields

inviter    :: D1zk0rd.User
uses       :: Int
max_uses   :: Int
max_age    :: Int
temporary  :: Bool
created_at :: DateTime
revoked    :: Bool

D1zk0rd.Member — Type

A Guild member. More details here.

Fields

user          :: Optional{D1zk0rd.User}
nick          :: OptionalNullable{String}
roles         :: Vector{Snowflake}
joined_at     :: DateTime
premium_since :: OptionalNullable{DateTime}
deaf          :: Optional{Bool}
mute          :: Optional{Bool}

D1zk0rd.MessageActivity — Type

A Message activity. More details here.

Fields

type     :: Int
party_id :: Optional{String}

D1zk0rd.MessageApplication — Type

A Rich Presence Message's application information. More details here.

Fields

id          :: Snowflake
cover_image :: Optional{String}
description :: String
icon        :: String
name        :: String

D1zk0rd.MessageType — Module

A Message's type. More details here.

D1zk0rd.MessageActivityType — Module

A Message's activity type. More details here.

D1zk0rd.Overwrite — Type

A permission overwrite. More details here.

Fields

id    :: Snowflake
type  :: Int
allow :: String
deny  :: String

D1zk0rd.Presence — Type

A User's presence. More details here.

Fields

user       :: D1zk0rd.User
roles      :: Optional{Vector{Snowflake}}
game       :: Nullable{D1zk0rd.Activity}
guild_id   :: Optional{Snowflake}
status     :: String
activities :: Vector{D1zk0rd.Activity}

D1zk0rd.Reaction — Type

A Message reaction. More details here.

Fields

count :: Int
me    :: Bool
emoji :: D1zk0rd.Emoji

D1zk0rd.Role — Type

A User role. More details here.

Fields

id          :: Snowflake
name        :: String
color       :: Optional{Int}
hoist       :: Optional{Bool}
position    :: Optional{Int}
permissions :: Optional{String}
managed     :: Optional{Bool}
mentionable :: Optional{Bool}

D1zk0rd.SelectOption — Type

A select option. More details here.

Fields

label       :: String
value       :: String
description :: Optional{String}
emoji       :: Optional{D1zk0rd.Emoji}
default     :: Optional{Bool}

D1zk0rd.User — Type

A Discord user. More details here.

Fields

id            :: Snowflake
username      :: Optional{String}
discriminator :: Optional{String}
avatar        :: OptionalNullable{String}
bot           :: Optional{Bool}
mfa_enabled   :: Optional{Bool}
locale        :: Optional{String}
verified      :: Optional{Bool}
email         :: OptionalNullable{String}

D1zk0rd.VoiceRegion — Type

A region for a Guild's voice server. More details here.

Fields

id         :: String
name       :: String
vip        :: Bool
optimal    :: Bool
deprecated :: Bool
custom     :: Bool

D1zk0rd.VoiceState — Type

A User's voice connection status. More details here.

Fields

guild_id   :: Optional{Snowflake}
channel_id :: Nullable{Snowflake}
user_id    :: Snowflake
member     :: Optional{D1zk0rd.Member}
session_id :: String
deaf       :: Bool
mute       :: Bool
self_deaf  :: Bool
self_mute  :: Bool
suppress   :: Bool

D1zk0rd.Webhook — Type

A Webhook. More details here.

Fields

id         :: Snowflake
guild_id   :: Optional{Snowflake}
channel_id :: Snowflake
user       :: Optional{D1zk0rd.User}
name       :: Nullable{String}
avatar     :: Nullable{String}
token      :: Optional{String}