Events

Note that Snowflake === UInt64. Unions with Nothing indicate that a field is nullable, whereas Unions with Missing indicate that a field is optional. –- NOTE

The Event types are deprecated. They will be removed in 0.3.z.

Dizkord.AbstractEvent — Type

An incoming event sent over the gateway. Also a catch-all event: Handlers defined on this type will execute on all events.

Dizkord.FallbackEvent — Type

A type for defining handlers on any events which would otherwise have no non-default handler. Handlers for this type must accept an AbstractEvent.

Dizkord.UnknownEvent — Type

An unknown event. When an event can't be parsed, due to an unknown type or any other error, it will appear as an UnknownEvent. The fields follow the schema defined here.

Channels

Dizkord.ChannelCreate — Type

Sent when a new DiscordChannel is created.

Fields

channel :: Dizkord.DiscordChannel

Dizkord.ChannelUpdate — Type

Sent when a DiscordChannel is updated.

Fields

channel :: Dizkord.DiscordChannel

Dizkord.ChannelDelete — Type

Sent when a DiscordChannel is deleted.

Fields

channel :: Dizkord.DiscordChannel

Dizkord.ChannelPinsUpdate — Type

Sent when a DiscordChannel's pins are updated.

Fields

channel_id         :: Snowflake
last_pin_timestamp :: Nullable{DateTime}

Guilds

Dizkord.GuildCreate — Type

Sent when a new Guild is created.

Fields

guild :: Dizkord.Guild

Dizkord.GuildUpdate — Type

Sent when a Guild is updated.

Fields

guild :: Dizkord.Guild

Dizkord.GuildDelete — Type

Sent when a Guild is deleted.

Fields

guild :: Dizkord.AbstractGuild

Dizkord.GuildBanAdd — Type

Sent when a User is banned from a Guild.

Fields

guild_id :: Snowflake
user     :: Dizkord.User

Dizkord.GuildBanRemove — Type

Sent when a User is unbanned from a Guild.

Fields

guild_id :: Snowflake
user     :: Dizkord.User

Dizkord.GuildEmojisUpdate — Type

Sent when a Guild has its Emojis updated.

Fields

guild_id :: Snowflake
emojis   :: Vector{Dizkord.Emoji}

Dizkord.GuildIntegrationsUpdate — Type

Sent when a Guild has its Integrations updated.

Fields

guild_id :: Snowflake

Dizkord.GuildMemberAdd — Type

Sent when a Member is added to a Guild.

Fields

guild_id :: Snowflake
member   :: Dizkord.Member

Dizkord.GuildMemberRemove — Type

Sent when a Member is removed from a Guild.

Fields

guild_id :: Snowflake
user     :: Dizkord.User

Dizkord.GuildMemberUpdate — Type

Sent when a Member is updated in a Guild.

Fields

guild_id :: Snowflake
roles    :: Vector{Snowflake}
user     :: Dizkord.User
nick     :: Nullable{String}

Dizkord.GuildMembersChunk — Type

Sent when the Client requests guild members with request_guild_members.

Fields

guild_id :: Snowflake
members  :: Vector{Dizkord.Member}

Dizkord.GuildRoleCreate — Type

Sent when a new Role is created in a Guild.

Fields

guild_id :: Snowflake
role     :: Dizkord.Role

Dizkord.GuildRoleUpdate — Type

Sent when a Role is updated in a Guild.

Fields

guild_id :: Snowflake
role     :: Dizkord.Role

Dizkord.GuildRoleDelete — Type

Sent when a Role is deleted from a Guild.

Fields

guild_id :: Snowflake
role_id  :: Snowflake

Messages

Dizkord.MessageCreate — Type

Sent when a Message is sent.

Fields

message :: Dizkord.Message

Dizkord.MessageUpdate — Type

Sent when a Message is updated.

Fields

message :: Dizkord.Message

Dizkord.MessageDelete — Type

Sent when a Message is deleted.

Fields

id         :: Snowflake
channel_id :: Snowflake
guild_id   :: Optional{Snowflake}

Dizkord.MessageDeleteBulk — Type

Sent when multiple Messages are deleted in bulk.

Fields

ids        :: Vector{Snowflake}
channel_id :: Snowflake
guild_id   :: Optional{Snowflake}

Dizkord.MessageReactionAdd — Type

Sent when a Reaction is added to a Message.

Fields

user_id    :: Snowflake
channel_id :: Snowflake
message_id :: Snowflake
guild_id   :: Optional{Snowflake}
emoji      :: Dizkord.Emoji

Dizkord.MessageReactionRemove — Type

Sent when a Reaction is removed from a Message.

Fields

user_id    :: Snowflake
channel_id :: Snowflake
message_id :: Snowflake
guild_id   :: Optional{Snowflake}
emoji      :: Dizkord.Emoji

Dizkord.MessageReactionRemoveAll — Type

Sent when all Reactions are removed from a Message.

Fields

channel_id :: Snowflake
message_id :: Snowflake
guild_id   :: Optional{Snowflake}

Presence

Dizkord.PresenceUpdate — Type

Sent when a User's Presence is updated.

Fields

presence :: Dizkord.Presence

Dizkord.TypingStart — Type

Sent when a User begins typing.

Fields

channel_id :: Snowflake
guild_id   :: Optional{Snowflake}
user_id    :: Snowflake
timestamp  :: DateTime

Dizkord.UserUpdate — Type

Sent when a User's details are updated.

Fields

user :: Dizkord.User

Voice

Dizkord.VoiceStateUpdate — Type

Sent when a User updates their VoiceState.

Fields

state :: Dizkord.VoiceState

Dizkord.VoiceServerUpdate — Type

Sent when a Guild's voice server is updated.

Fields

token    :: String
guild_id :: Snowflake
endpoint :: String

Webhooks

Dizkord.WebhooksUpdate — Type

Sent when a DiscordChannel's Webhooks are updated.

Fields

guild_id   :: Snowflake
channel_id :: Snowflake

Connecting

Dizkord.Ready — Type

Sent when the Client has authenticated, and contains the initial state.

Fields

v                :: Int
user             :: Dizkord.User
private_channels :: Vector{Dizkord.DiscordChannel}
guilds           :: Vector{Dizkord.UnavailableGuild}
session_id       :: String
_trace           :: Vector{String}

Dizkord.Resumed — Type

Sent when a Client resumes its connection.

Fields

_trace :: Vector{String}

Handlers

Handlers are mirrors of Events prefixed with On. They always contain a function that takes a Context as an argument. Each handler has an analog context, that context is simply the handler suffixed with Context. For example the MessageCreate event has a handler called OnMessageCreate and a context called [OnMessageCreateContext].

Channels

Missing docstring.

Missing docstring for ChannelCreate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChannelUpdate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChannelDelete. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChannelPinsUpdate. Check Documenter's build log for details.

Guilds

Dizkord.OnGuildCreate — Type

Handler for GuildCreate event

Fields

f :: Function

Dizkord.OnGuildUpdate — Type

Handler for GuildUpdate event

Fields

f :: Function

Dizkord.OnGuildDelete — Type

Handler for GuildDelete event

Messages

Missing docstring.

Missing docstring for MessageCreate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MessageUpdate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MessageDelete. Check Documenter's build log for details.

Connecting

Missing docstring.

Missing docstring for Ready. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Resumed. Check Documenter's build log for details.