Events

Events are how Discord communicates and interacts with our application. In Dizkord.jl, we use the Handler type in conjunction with the Context type to handle them.

Handler

A Handler is simple to register for an event. To register a handler for the Ready event:

# `c` is a client generated previously.
h = OnReady() do (ctx)
    println("I'm ready!")
end

add_handler!(c, h)

Note that as OnReady is a common event, there is a convenience method for it. As such the following code is fully equivalent:

on_ready!(c) do (ctx)
    println("I'm ready!")
end

(on_ready! creates an OnReady handler and adds it to the client.) You can find a list of all gateway events here. Simply add On before their NAME to get the name of the associated handler!

Dizkord.Handler — Type

Handler(
    f::Function
    d::Dict{Symbol, Any}
)

Handler is a wrapper for a Dict{Symbol, Any} that also contains a function.

Context

The context is simply the payload received from the interaction. Exceptions:

The MessageCreate context contains a Message.

The OnGuildCreate and OnGuildUpdate contexts contain a Guild.

The OnInteractionCreate context contains an Interaction.

You can find the expected payloads for events here.

Dizkord.Context — Type

Context(
    data::Dict{Symbol, Any}
)

Context is a wrapper for a Dict{Symbol, Any} with some special functionality.

Special Handlers

Dizkord.on_message! — Function

on_message!(
    f::Function
    c::Client
)

Adds a handler for the MESSAGE_CREATE gateway event. The f parameter's signature should be:

    (ctx::Context) -> Any 

Dizkord.on_ready! — Function

on_ready!(
    f::Function
    c::Client
)

Adds a handler for the READY gateway event. The f parameter signature should be:

    (ctx::Context) -> Any 

Dizkord.command! — Function

command!(
    f::Function
    c::Client
    name::AbstractString
    description::AbstractString;
    kwargs...
)

Adds a handler for INTERACTION CREATE gateway events where the InteractionData's name field matches name. Adds this command to c.commands or c.guild_commands based on the presence of guild. The f parameter signature should be:

    (ctx::Context, args...) -> Any 

Where args is a list of all the Command Options

For example a command that takes a user u and a number n as input should have this signature:

(ctx::Context, u::User, n::Int) -> Any

and the arguments would automatically get converted.

Note: The argument names must match the Option names. The arguments can be ordered in any way. If no type is specified, no conversion will be performed, so Discord objects will be Snowflakes.