Types

Ekztazy.Message — Type

A message sent to a DiscordChannel. More details here.

Fields

id                :: Snowflake
channel_id        :: Snowflake
guild_id          :: Optional{Snowflake}
author            :: Optional{Ekztazy.User}
member            :: Optional{Ekztazy.Member}
content           :: Optional{String}
timestamp         :: Optional{DateTime}
edited_timestamp  :: OptionalNullable{DateTime}
tts               :: Optional{Bool}
mention_everyone  :: Optional{Bool}
mentions          :: Optional{Vector{Ekztazy.User}}
mention_roles     :: Optional{Vector{Snowflake}}
attachments       :: Optional{Vector{Ekztazy.Attachment}}
message_reference :: Optional{Ekztazy.MessageReference}
embeds            :: Optional{Vector{Ekztazy.Embed}}
reactions         :: Optional{Vector{Ekztazy.Reaction}}
nonce             :: OptionalNullable{Snowflake}
pinned            :: Optional{Bool}
webhook_id        :: Optional{Snowflake}
type              :: Optional{Int}
activity          :: Optional{Ekztazy.MessageActivity}
application       :: Optional{Ekztazy.MessageApplication}

Ekztazy.AbstractGuild — Type

A Discord guild (server). Can either be an UnavailableGuild or a Guild.

Ekztazy.Guild — Type

A Discord guild (server). More details here.

Fields

id                            :: Snowflake
name                          :: String
icon                          :: Nullable{String}
splash                        :: OptionalNullable{String}
owner                         :: Optional{Bool}
owner_id                      :: Optional{Snowflake}
permissions                   :: Optional{String}
region                        :: Optional{String}
afk_channel_id                :: OptionalNullable{Snowflake}
afk_timeout                   :: Optional{Int}
embed_enabled                 :: Optional{Bool}
embed_channel_id              :: OptionalNullable{Snowflake}
verification_level            :: Optional{Int}
default_message_notifications :: Optional{Int}
explicit_content_filter       :: Optional{Int}
roles                         :: Optional{Vector{Ekztazy.Role}}
emojis                        :: Optional{Vector{Ekztazy.Emoji}}
features                      :: Optional{Vector{String}}
mfa_level                     :: Optional{Int}
application_id                :: OptionalNullable{Snowflake}
widget_enabled                :: Optional{Bool}
widget_channel_id             :: OptionalNullable{Snowflake}
system_channel_id             :: OptionalNullable{Snowflake}
joined_at                     :: Optional{DateTime}
large                         :: Optional{Bool}
unavailable                   :: Optional{Bool}
member_count                  :: Optional{Int}
max_members                   :: Optional{Int}
voice_states                  :: Optional{Vector{Ekztazy.VoiceState}}
members                       :: Optional{Vector{Ekztazy.Member}}
channels                      :: Optional{Vector{Ekztazy.DiscordChannel}}
presences                     :: Optional{Vector{Ekztazy.Presence}}
max_presences                 :: OptionalNullable{Int}
vanity_url_code               :: OptionalNullable{String}
description                   :: OptionalNullable{String}
banner                        :: OptionalNullable{String}

Ekztazy.Activity — Type

A User activity. More details here.

Fields

name           :: String
type           :: Int
url            :: OptionalNullable{String}
timestamps     :: Optional{Ekztazy.ActivityTimestamps}
application_id :: Optional{Snowflake}
details        :: OptionalNullable{String}
state          :: OptionalNullable{String}
emoji          :: OptionalNullable{Ekztazy.ActivityEmoji}
party          :: Optional{Ekztazy.ActivityParty}
assets         :: Optional{Ekztazy.ActivityAssets}
secrets        :: Optional{Ekztazy.ActivitySecrets}
instance       :: Optional{Bool}
flags          :: Optional{Int}

Ekztazy.ActivityTimestamps — Type

The start and stop times of an Activity. More details here.

Fields

start :: Optional{DateTime}
stop  :: Optional{DateTime}

Ekztazy.ActivityParty — Type

The current party of an Activity's player. More details here.

Fields

id   :: Optional{String}
size :: Optional{Vector{Int}}

Ekztazy.ActivityAssets — Type

Images and hover text for an Activity. More details here.

Fields

large_image :: Optional{String}
large_text  :: Optional{String}
small_image :: Optional{String}
small_text  :: Optional{String}

Ekztazy.ActivitySecrets — Type

Secrets for Rich Presence joining and spectating of an Activity. More details here.

Fields

join     :: Optional{String}
spectate :: Optional{String}
match    :: Optional{String}

Ekztazy.ActivityType — Module

An Activity's type. Available values are GAME, STREAMING, LISTENING, WATCHING, and COMPETING. More details here.

Ekztazy.ActivityFlags — Module

Flags which indicate what an Activity payload contains. More details here.

Ekztazy.ApplicationCommand — Type

An Application Command. More details here.

Fields

id                  :: OptionalNullable{Snowflake}
type                :: Optional{Int}
application_id      :: Snowflake
guild_id            :: Optional{Snowflake}
name                :: String
description         :: String
options             :: Optional{Vector{Ekztazy.ApplicationCommandOption}}
default_permissions :: Optional{Bool}
version             :: Optional{Snowflake}

Ekztazy.ApplicationCommandOption — Type

Application Command Option. More details here.

Fields

value         :: Any
type          :: Optional{Int}
name          :: Optional{String}
description   :: Optional{String}
required      :: Optional{Bool}
min_value     :: Optional{Number}
max_value     :: Optional{Number}
autocomplete  :: Optional{Bool}
choices       :: Optional{Vector{Ekztazy.ApplicationCommandChoice}}
options       :: Optional{Vector{Ekztazy.ApplicationCommandOption}}
channel_types :: Optional{Vector{Int}}
focused       :: Optional{Bool}

Ekztazy.ApplicationCommandChoice — Type

Application Command Choice. More details here.

Fields

name  :: String
value :: Union{Number, String}

Ekztazy.Attachment — Type

A Message attachment. More details here.

Fields

id        :: Snowflake
filename  :: String
size      :: Int
url       :: String
proxy_url :: String
height    :: Optional{Int}
width     :: Optional{Int}

Ekztazy.AuditLog — Type

An audit log. More details here.

Fields

webhooks          :: Vector{Ekztazy.Webhook}
users             :: Vector{Ekztazy.User}
audit_log_entries :: Vector{Ekztazy.AuditLogEntry}

Ekztazy.AuditLogEntry — Type

An entry in an AuditLog. More details here.

Fields

target_id   :: Nullable{Snowflake}
changes     :: Optional{Vector{Ekztazy.AuditLogChange}}
user_id     :: Snowflake
id          :: Snowflake
action_type :: Int
options     :: Optional{Ekztazy.AuditLogOptions}
reason      :: Optional{String}

Ekztazy.AuditLogChange — Type

A change item in an AuditLogEntry.

The first type parameter is the type of new_value and old_value. The second is the type of the entity that new_value and old_value belong(ed) to.

More details here.

Fields

new_value :: Optional{T} where T
old_value :: Optional{T} where T
key       :: String
type      :: Type{U} where U

Ekztazy.AuditLogOptions — Type

Optional information in an AuditLogEntry. More details here.

Fields

delete_member_days :: Optional{Int}
members_removed    :: Optional{Int}
channel_id         :: Optional{Snowflake}
count              :: Optional{Int}
id                 :: Optional{Snowflake}
type               :: Optional{Int}
role_name          :: Optional{String}

Ekztazy.ActionType — Module

AuditLog action types. More details here.

Ekztazy.Ban — Type

A User ban. More details here.

Fields

reason :: Nullable{String}
user   :: Ekztazy.User

Ekztazy.DiscordChannel — Type

A Discord channel. More details here.

Note: The name Channel is already used, hence the prefix.

Fields

id                    :: Snowflake
type                  :: Optional{Int}
guild_id              :: Optional{Snowflake}
position              :: Optional{Int}
permission_overwrites :: Optional{Vector{Ekztazy.Overwrite}}
name                  :: Optional{String}
topic                 :: OptionalNullable{String}
nsfw                  :: Optional{Bool}
last_message_id       :: OptionalNullable{Snowflake}
bitrate               :: Optional{Int}
user_limit            :: Optional{Int}
rate_limit_per_user   :: Optional{Int}
recipients            :: Optional{Vector{Ekztazy.User}}
icon                  :: OptionalNullable{String}
owner_id              :: Optional{Snowflake}
application_id        :: Optional{Snowflake}
parent_id             :: OptionalNullable{Snowflake}
last_pin_timestamp    :: OptionalNullable{DateTime}

Ekztazy.Connection — Type

A User connection to an external service (Twitch, YouTube, etc.). More details here.

Fields

id           :: String
name         :: String
type         :: String
revoked      :: Bool
integrations :: Vector{Ekztazy.Integration}

Ekztazy.Component — Type

An interactable component. More details here.

Fields

type        :: Int
custom_id   :: Optional{String}
value       :: Optional{String}
disabled    :: Optional{Bool}
style       :: Optional{Int}
label       :: Optional{String}
emoji       :: Optional{Ekztazy.Emoji}
url         :: Optional{String}
options     :: Optional{Vector{Ekztazy.SelectOption}}
placeholder :: Optional{String}
min_values  :: Optional{Int}
max_values  :: Optional{Int}
components  :: Optional{Vector{Ekztazy.Component}}

Ekztazy.Embed — Type

A Message embed. More details here.

Fields

title       :: Optional{String}
type        :: Optional{String}
description :: Optional{String}
url         :: Optional{String}
timestamp   :: Optional{DateTime}
color       :: Optional{Int}
footer      :: Optional{Ekztazy.EmbedFooter}
image       :: Optional{Ekztazy.EmbedImage}
thumbnail   :: Optional{Ekztazy.EmbedThumbnail}
video       :: Optional{Ekztazy.EmbedVideo}
provider    :: Optional{Ekztazy.EmbedProvider}
author      :: Optional{Ekztazy.EmbedAuthor}
fields      :: Optional{Vector{Ekztazy.EmbedField}}

Ekztazy.EmbedThumbnail — Type

An Embed's thumbnail image information. More details here.

Fields

url       :: Optional{String}
proxy_url :: Optional{String}
height    :: Optional{Int}
width     :: Optional{Int}

Ekztazy.EmbedVideo — Type

An Embed's video information. More details here.

Fields

url    :: Optional{String}
height :: Optional{Int}
width  :: Optional{Int}

Ekztazy.EmbedImage — Type

An Embed's image information. More details here.

Fields

url       :: Optional{String}
proxy_url :: Optional{String}
height    :: Optional{Int}
width     :: Optional{Int}

Ekztazy.EmbedProvider — Type

An Embed's provider information. More details here.

Fields

name :: Optional{String}
url  :: OptionalNullable{String}

Ekztazy.EmbedAuthor — Type

An Embed's author information. More details here.

Fields

name           :: Optional{String}
url            :: Optional{String}
icon_url       :: Optional{String}
proxy_icon_url :: Optional{String}

Ekztazy.EmbedFooter — Type

An Embed's footer information. More details here.

Fields

text           :: String
icon_url       :: Optional{String}
proxy_icon_url :: Optional{String}

Ekztazy.EmbedField — Type

An Embed field. More details here.

Fields

name   :: String
value  :: String
inline :: Optional{Bool}

Ekztazy.Emoji — Type

An emoji. More details here.

Fields

id             :: Nullable{Snowflake}
name           :: String
roles          :: Optional{Vector{Snowflake}}
user           :: Optional{Ekztazy.User}
require_colons :: Optional{Bool}
managed        :: Optional{Bool}
animated       :: Optional{Bool}

Ekztazy.UnavailableGuild — Type

An unavailable Discord guild (server). More details here.

Fields

id          :: Snowflake
unavailable :: Optional{Bool}

Ekztazy.VerificationLevel — Module

A Guild's verification level. More details here.

Ekztazy.MessageNotificationLevel — Module

A Guild's default message notification level. More details here.

Ekztazy.ExplicitContentFilterLevel — Module

A Guild's explicit content filter level. More details here.

Ekztazy.MFALevel — Module

A Guild's MFA level. More details here.

Ekztazy.GuildEmbed — Type

A Guild embed. More details here.

Fields

enabled    :: Bool
channel_id :: Nullable{Snowflake}

Ekztazy.Integration — Type

A Guild integration. More details here.

Fields

id                  :: Snowflake
name                :: String
type                :: String
enabled             :: Bool
syncing             :: Bool
role_id             :: Snowflake
expire_behaviour    :: Int
expire_grace_period :: Int
user                :: Ekztazy.User
account             :: Ekztazy.IntegrationAccount
synced_at           :: DateTime

Ekztazy.IntegrationAccount — Type

An Integration account. More details here.

Fields

id   :: String
name :: String

Ekztazy.Interaction — Type

An interaction. More details here.

Fields

id             :: Nullable{Snowflake}
application_id :: Nullable{Snowflake}
type           :: Int
data           :: OptionalNullable{Ekztazy.InteractionData}
guild_id       :: Optional{Snowflake}
channel_id     :: Optional{Snowflake}
member         :: Optional{Ekztazy.Member}
user           :: Optional{Ekztazy.User}
token          :: String
version        :: Optional{Int}
message        :: Optional{Ekztazy.Message}

Ekztazy.InteractionData — Type

Data for an interaction. More details here.

Fields

id             :: OptionalNullable{Snowflake}
name           :: OptionalNullable{String}
type           :: OptionalNullable{Int}
resolved       :: Optional{Ekztazy.ResolvedData}
options        :: Optional{Vector{Ekztazy.ApplicationCommandOption}}
custom_id      :: OptionalNullable{String}
component_type :: OptionalNullable{Int}
components     :: OptionalNullable{Vector{Ekztazy.Component}}
values         :: Optional{Vector{String}}
target_id      :: Optional{Snowflake}

Ekztazy.Invite — Type

An invite to a Guild. More details here.

Fields

code                       :: String
guild                      :: Optional{Ekztazy.Guild}
channel                    :: Ekztazy.DiscordChannel
approximate_presence_cound :: Optional{Int}
approximate_member_count   :: Optional{Int}

Ekztazy.InviteMetadata — Type

Metadata for an Invite. More details here.

Fields

inviter    :: Ekztazy.User
uses       :: Int
max_uses   :: Int
max_age    :: Int
temporary  :: Bool
created_at :: DateTime
revoked    :: Bool

Ekztazy.Member — Type

A Guild member. More details here.

Fields

user          :: Optional{Ekztazy.User}
nick          :: OptionalNullable{String}
roles         :: Vector{Snowflake}
joined_at     :: DateTime
premium_since :: OptionalNullable{DateTime}
deaf          :: Optional{Bool}
mute          :: Optional{Bool}

Ekztazy.MessageActivity — Type

A Message activity. More details here.

Fields

type     :: Int
party_id :: Optional{String}

Ekztazy.MessageApplication — Type

A Rich Presence Message's application information. More details here.

Fields

id          :: Snowflake
cover_image :: Optional{String}
description :: String
icon        :: String
name        :: String

Ekztazy.MessageType — Module

A Message's type. More details here.

Ekztazy.MessageActivityType — Module

A Message's activity type. More details here.

Ekztazy.Overwrite — Type

A permission overwrite. More details here.

Fields

id    :: Snowflake
type  :: Int
allow :: String
deny  :: String

Ekztazy.Presence — Type

A User's presence. More details here.

Fields

user       :: Ekztazy.User
roles      :: Optional{Vector{Snowflake}}
game       :: Nullable{Ekztazy.Activity}
guild_id   :: Optional{Snowflake}
status     :: String
activities :: Vector{Ekztazy.Activity}

Ekztazy.Reaction — Type

A Message reaction. More details here.

Fields

count :: Int
me    :: Bool
emoji :: Ekztazy.Emoji

Ekztazy.Role — Type

A User role. More details here.

Fields

id          :: Snowflake
name        :: String
color       :: Optional{Int}
hoist       :: Optional{Bool}
position    :: Optional{Int}
permissions :: Optional{String}
managed     :: Optional{Bool}
mentionable :: Optional{Bool}

Ekztazy.SelectOption — Type

A select option. More details here.

Fields

label       :: String
value       :: String
description :: Optional{String}
emoji       :: Optional{Ekztazy.Emoji}
default     :: Optional{Bool}

Ekztazy.User — Type

A Discord user. More details here.

Fields

id            :: Snowflake
username      :: Optional{String}
discriminator :: Optional{String}
avatar        :: OptionalNullable{String}
bot           :: Optional{Bool}
mfa_enabled   :: Optional{Bool}
locale        :: Optional{String}
verified      :: Optional{Bool}
email         :: OptionalNullable{String}

Ekztazy.VoiceRegion — Type

A region for a Guild's voice server. More details here.

Fields

id         :: String
name       :: String
vip        :: Bool
optimal    :: Bool
deprecated :: Bool
custom     :: Bool

Ekztazy.VoiceState — Type

A User's voice connection status. More details here.

Fields

guild_id   :: Optional{Snowflake}
channel_id :: Nullable{Snowflake}
user_id    :: Snowflake
member     :: Optional{Ekztazy.Member}
session_id :: String
deaf       :: Bool
mute       :: Bool
self_deaf  :: Bool
self_mute  :: Bool
suppress   :: Bool

Ekztazy.Webhook — Type

A Webhook. More details here.

Fields

id         :: Snowflake
guild_id   :: Optional{Snowflake}
channel_id :: Snowflake
user       :: Optional{Ekztazy.User}
name       :: Nullable{String}
avatar     :: Nullable{String}
token      :: Optional{String}